AtmosModel

AtmosProblem

ClimateMachine.Atmos.AtmosProblem — Type

AtmosProblem

The default problem definition (initial and boundary conditions) for AtmosModel.

AtmosModel balance law

ClimateMachine.Atmos.AtmosModel — Type

AtmosModel <: BalanceLaw

A BalanceLaw for atmosphere modeling. Users may over-ride prescribed default values for each field.

Usage

AtmosModel(
    param_set,
    problem,
    orientation,
    ref_state,
    turbulence,
    hyperdiffusion,
    spongelayer,
    moisture,
    precipitation,
    radiation,
    source,
    tracers,
    data_config,
)

Fields

param_set

Parameter Set (type to dispatch on, e.g., planet parameters. See CLIMAParameters.jl package)

problem

Problem (initial and boundary conditions)

orientation

An orientation model

ref_state

Reference State (For initial conditions, or for linearisation when using implicit solvers)

turbulence

Turbulence Closure (Equations for dynamics of under-resolved turbulent flows)

turbconv

Turbulence Convection Closure (e.g., EDMF)

hyperdiffusion

Hyperdiffusion Model (Equations for dynamics of high-order spatial wave attenuation)

viscoussponge

Viscous sponge layers

moisture

Moisture Model (Equations for dynamics of moist variables)

precipitation

Precipitation Model (Equations for dynamics of precipitating species)

radiation

Radiation Model (Equations for radiative fluxes)

source

Source Terms (Problem specific source terms)

tracers

Tracer Terms (Equations for dynamics of active and passive tracers)

lsforcing

Large-scale forcing (Forcing information from GCMs, reanalyses, or observations)

data_config

Data Configuration (Helper field for experiment configuration)

AtmosModel methods

ClimateMachine.BalanceLaws.flux_first_order! — Method

flux_first_order!(
    m::AtmosModel,
    flux::Grad,
    state::Vars,
    aux::Vars,
    t::Real
)

Computes and assembles non-diffusive fluxes in the model equations.

ClimateMachine.BalanceLaws.flux_second_order! — Method

flux_second_order!(
    atmos::AtmosModel,
    flux::Grad,
    state::Vars,
    diffusive::Vars,
    hyperdiffusive::Vars,
    aux::Vars,
    t::Real
)

Diffusive fluxes in AtmosModel. Viscosity, diffusivity are calculated in the turbulence subcomponent and accessed within the diffusive flux function. Contributions from subcomponents are then assembled (pointwise).

ClimateMachine.BalanceLaws.init_state_auxiliary! — Method

init_state_auxiliary!(
    m::AtmosModel,
    aux::Vars,
    grid,
    direction
)

Initialise auxiliary variables for each AtmosModel subcomponent. Store Cartesian coordinate information in aux.coord.

ClimateMachine.BalanceLaws.source! — Method

source!(
    m::AtmosModel,
    source::Vars,
    state::Vars,
    diffusive::Vars,
    aux::Vars,
    t::Real,
    direction::Direction,
)

Computes (and assembles) source terms S(Y) in:

∂Y
-- = - ∇ • F + S(Y)
∂t

ClimateMachine.BalanceLaws.init_state_prognostic! — Method

init_state_prognostic!(
    m::AtmosModel,
    state::Vars,
    aux::Vars,
    localgeo,
    t,
    args...,
)

Initialise state variables. args... provides an option to include configuration data (current use cases include problem constants, spline-interpolants).

Reference states

ClimateMachine.Atmos.HydrostaticState — Type

HydrostaticState{P,T} <: ReferenceState

A hydrostatic state specified by a virtual temperature profile and relative humidity.

By default, this is a dry hydrostatic reference state.

ClimateMachine.Atmos.InitStateBC — Type

InitStateBC

Set the value at the boundary to match the init_state_prognostic! function. This is mainly useful for cases where the problem has an explicit solution.

TODO: This should be fixed later once BCs are figured out (likely want

different things here?)

ClimateMachine.Atmos.ReferenceState — Type

ReferenceState

Hydrostatic reference state, for example, used as initial condition or for linearization.

ClimateMachine.Atmos.NoReferenceState — Type

NoReferenceState <: ReferenceState

No reference state used

Thermodynamics

ClimateMachine.Atmos.recover_thermo_state — Function

recover_thermo_state(atmos::AtmosModel, state::Vars, aux::Vars)

An atmospheric thermodynamic state.

Warn

While recoverthermostate is an ideal long-term solution, right now we are directly calling newthermostate to avoid inconsistent aux states in kernels where the aux states are out of sync with the boundary state.

TODO:

- Allow a safe way to call
`recover_thermo_state(state, moist::EquilMoist, ...)`

ClimateMachine.Atmos.new_thermo_state — Function

new_thermo_state(atmos::AtmosModel, state::Vars, aux::Vars)

Create a new thermodynamic state, based on the state, and not the aux state.

Note

This method calls the iterative saturation adjustment procedure for EquilMoist models.

Moisture and Precipitation

ClimateMachine.Atmos.DryModel — Type

DryModel

Assumes the moisture components is in the dry limit.

ClimateMachine.Atmos.EquilMoist — Type

EquilMoist

Assumes the moisture components are computed via thermodynamic equilibrium.

ClimateMachine.Atmos.NonEquilMoist — Type

NonEquilMoist

Does not assume that the moisture components are in equilibrium.

ClimateMachine.Atmos.NoPrecipitation — Type

NoPrecipitation <: PrecipitationModel

No precipitation.

ClimateMachine.Atmos.RainModel — Type

RainModel <: PrecipitationModel

Precipitation model with rain.

ClimateMachine.Atmos.RainSnowModel — Type

RainSnowModel <: PrecipitationModel

Precipitation model with rain and snow.

Stabilization

ClimateMachine.Atmos.RayleighSponge — Type

RayleighSponge{PV <: Momentum, FT} <: TendencyDef{Source, PV}

Rayleigh Damping (Linear Relaxation) for top wall momentum components Assumes laterally periodic boundary conditions for LES flows. Momentum components are relaxed to reference values (zero velocities) at the top boundary.

BCs

ClimateMachine.Atmos.AtmosBC — Type

AtmosBC(momentum = Impenetrable(FreeSlip())
        energy   = Insulating()
        moisture = Impermeable()
        precipitation = OutflowPrecipitation()
        tracer  = ImpermeableTracer())

The standard boundary condition for AtmosModel. The default options imply a "no flux" boundary condition.

ClimateMachine.Atmos.DragLaw — Type

DragLaw(fn) :: MomentumDragBC

Drag law for momentum parallel to the boundary. The drag coefficient is C = fn(state, aux, t, normu_int_tan), where normu_int_tan is the internal speed parallel to the boundary. _int refers to the first interior node.

ClimateMachine.Atmos.Impermeable — Type

Impermeable() :: MoistureBC

No moisture flux.

ClimateMachine.Atmos.PrescribedMoistureFlux — Type

PrescribedMoistureFlux(fn) :: MoistureBC

Prescribe the net inward moisture flux across the boundary by fn, a function with signature fn(state, aux, t), returning the flux (in kg/m^2).

ClimateMachine.Atmos.BulkFormulaMoisture — Type

BulkFormulaMoisture(fn) :: MoistureBC

Calculate the net inward moisture flux across the boundary using the bulk formula. The drag coefficient is C_q = fn_C_q(state, aux, t, normu_int_tan). The surface qtot at the boundary is `qtot = fnqtot(state, aux, t)`.

Return the flux (in kg m^-2 s^-1).

ClimateMachine.Atmos.FreeSlip — Type

FreeSlip() :: MomentumDragBC

No surface drag on momentum parallel to the boundary.

ClimateMachine.Atmos.PrescribedTemperature — Type

PrescribedTemperature(fn) :: EnergyBC

Prescribe the temperature at the boundary by fn, a function with signature fn(state, aux, t) returning the temperature (in K).

ClimateMachine.Atmos.PrescribedEnergyFlux — Type

PrescribedEnergyFlux(fn) :: EnergyBC

Prescribe the net inward energy flux across the boundary by fn, a function with signature fn(state, aux, t), returning the flux (in W/m^2).

ClimateMachine.Atmos.BulkFormulaEnergy — Type

BulkFormulaEnergy(fn) :: EnergyBC

Calculate the net inward energy flux across the boundary. The drag coefficient is C_h = fn_C_h(atmos, state, aux, t, normu_int_tan). The surface temperature and qtot are `T, qtot = fnTandqtot(atmos, state, aux, t)`.

Return the flux (in W m^-2).

ClimateMachine.Atmos.OutflowPrecipitation — Type

OutflowPrecipitation() :: PrecipitationBC

Free flux out of the domain.

ClimateMachine.Atmos.ImpermeableTracer — Type

ImpermeableTracer() :: TracerBC

No tracer diffusion across boundary

ClimateMachine.Atmos.Impenetrable — Type

Impenetrable(drag::MomentumDragBC) :: MomentumBC

Defines an impenetrable wall model for momentum. This implies:

no flow in the direction normal to the boundary, and

flow parallel to the boundary is subject to the drag condition.

ClimateMachine.Atmos.Insulating — Type

Insulating() :: EnergyBC

No energy flux across the boundary.

ClimateMachine.Atmos.NoSlip — Type

NoSlip() :: MomentumDragBC

Zero momentum at the boundary.

ClimateMachine.Atmos.average_density — Function

average_density(ρ_sfc, ρ_int)

Average density between the surface and the interior point, given

ρ_sfc density at the surface

ρ_int density at the interior point

Sources

ClimateMachine.Atmos.RemovePrecipitation — Type

RemovePrecipitation{PV} <: TendencyDef{Source, PV}

A sink to q_tot when cloud condensate is exceeding a threshold. The threshold is defined either in terms of condensate or supersaturation. The removal rate is implemented as a relaxation term in the Microphysics_0M module. The default thresholds and timescale are defined in CLIMAParameters.jl.

ClimateMachine.Atmos.CreateClouds — Type

CreateClouds{PV <: Union{LiquidMoisture,IceMoisture}} <: TendencyDef{Source, PV}

A source/sink to q_liq and q_ice implemented as a relaxation towards equilibrium in the Microphysics module. The default relaxation timescales are defined in CLIMAParameters.jl.

ClimateMachine.Atmos.WarmRain_1M — Type

WarmRain_1M{FT} <: AbstractSource

A collection of source/sink terms related to 1-moment warm rain microphysics.

ClimateMachine.Atmos.RainSnow_1M — Type

RainSnow_1M{FT} <: AbstractSource

A collection of source/sink terms related to 1-moment rain and snow microphysics

Large-scale forcing

ClimateMachine.Atmos.NoLSForcing — Type

NoLSForcing <: LSForcingModel

No large-scale forcing

ClimateMachine.Atmos.HadGEMVertical — Type

Container for GCM variables from HadGEM2-A forcing,

used in the AMIP experiments