Microphysics

Methods

ClimateMachine.Microphysics.ζ_rai — Function

ζ_rai(param_set, ρ)

param_set - abstract set with earth parameters

ρ air density

Returns the proportionality coefficient between terminal velocity of an individual water drop and the square root of its radius.

ClimateMachine.Microphysics.n0_sno — Function

n0_sno(snow_param_set, q_sno, ρ)

snow_param_set - abstract set with snow microphysics parameters

q_sno - snow specific humidity

ρ - air density

Returns the intercept parameter of the assumed Marshal Palmer distribution of snow particles.

ClimateMachine.Microphysics.lambda — Function

lambda(q, ρ)

q - specific humidity of rain, ice or snow

ρ - air density

n0 - size disyribution parameter

α, β - mass(radius) parameters

Returns the rate parameter of the assumed size distribution of particles (rain drops, ice crystals, snow crystals).

Missing docstring.

Missing docstring for unpack_parameters. Check Documenter's build log for details.

ClimateMachine.Microphysics.supersaturation — Function

supersaturation(param_set, q, ρ, T, Liquid())
supersaturation(param_set, q, ρ, T, Ice())

param_set - abstract set with earth parameters

q - phase partition

ρ - air density,

T - air temperature

Liquid(), Ice() - liquid or ice phase to dispatch over.

Returns supersaturation (qv/qv_sat -1) over water or ice.

ClimateMachine.Microphysics.G_func — Function

G_func(param_set, T, Liquid())
G_func(param_set, T, Ice())

param_set - abstract set with earth parameters

T - air temperature

Liquid(), Ice() - liquid or ice phase to dispatch over.

Utility function combining thermal conductivity and vapor diffusivity effects.

ClimateMachine.Microphysics.terminal_velocity — Function

terminal_velocity(param_set, microphysics_param_set, ρ, q_)

param_set - abstract set with earth parameters

microphysics_param_set - abstract set with rain or snow microphysics parameters

ρ - air density

q_ - rain or snow specific humidity

Returns the mass weighted average terminal velocity assuming Marshall Palmer 1948 distribution of rain drops and snow crystals.

ClimateMachine.Microphysics.conv_q_vap_to_q_liq_ice — Function

conv_q_vap_to_q_liq_ice(cloud_param_set::ACPS, q_sat, q)
conv_q_vap_to_q_liq_ice(ice_param_set::AIPS, q_sat, q)

cloud_param_set - abstract set with cloud microphysics parameters

ice_param_set - abstract set with ice microphysics parameters

q_sat - PhasePartition at equilibrium

q - current PhasePartition

Returns the cloud water tendency due to condensation/evaporation or cloud ice tendency due to sublimation/resublimation. The tendency is obtained assuming a relaxation to equilibrium with constant timescale.

ClimateMachine.Microphysics.conv_q_liq_to_q_rai — Function

conv_q_liq_to_q_rai(rain_param_set, q_liq)

rain_param_set - abstract set with rain microphysics parameters

q_liq - liquid water specific humidity

Returns the q_rai tendency due to collisions between cloud droplets (autoconversion) parametrized following Kessler 1995.

ClimateMachine.Microphysics.conv_q_ice_to_q_sno — Function

conv_q_ice_to_q_sno(param_set, ice_param_set, q, ρ)

param_set - abstract set with earth parameters

ice_param_set - abstract set with ice microphysics parameters

q - phase partition

ρ - air density

T - air temperature

Returns the q_sno tendency due to autoconversion from ice. Parameterized following Harrington et al 1996 and Kaul et al 2015

ClimateMachine.Microphysics.accretion — Function

accretion(param_set, suspended_param_set, falling_param_set, q_susp, q_fall, ρ)

param_set - abstract set with earth parameters

suspended_param_set - abstract set with cloud microphysics or cloud ice parameters

falling_param_set - abstract set with rain or snow microphysics parameters

q_susp - cloud water or cloud ice specific humidity

q_fall - rain water or snow specific humidity

ρ - rain water or snow specific humidity

Returns the sink to suspended water (cloud water or cloud ice) due to collisions with falling water (rain or snow).

ClimateMachine.Microphysics.accretion_rain_sink — Function

accretion_rain_sink(param_set, ice_param_set, rain_param_set, q_ice, q_rai, ρ)

param_set - abstract set with earth parameters

ice_param_set - abstract set with ice microphysics parameters

rain_param_set - abstract set with rain microphysics parameters

q_ice - ice water specific humidity

q_rai - rain water specific humidity

ρ - air density

Returns the sink of rain water (partial source of snow) due to collisions with cloud ice.

ClimateMachine.Microphysics.accretion_snow_rain — Function

accretion_snow_rain(param_set, i_param_set, j_param_set, q_i, q_j, ρ)

i - snow for temperatures below freezing or rain for temperatures above freezing

j - rain for temperatures below freezing or rain for temperatures above freezing

param_set - abstract set with earth parameters

_param_set - abstract set with snow or rain microphysics parameters

q_ - specific humidity of snow or rain

ρ - air density

Returns the accretion rate between rain and snow. Collisions between rain and snow result in snow in temperatures below freezing andin rain in temperatures above freezing.

ClimateMachine.Microphysics.evaporation_sublimation — Function

evaporation_sublimation(param_set, rain_param_set, q_rai, ρ, T)
evaporation_sublimation(param_set, snow_param_set, q_sno, ρ, T)

param_set - abstract set with earth parameters

rain_param_set - abstract set with rain microphysics parameters

snow_param_set - abstract set with snow microphysics parameters

q - phase partition

q_rai - rain specific humidity

q_sno - snow specific humidity

ρ - air density

T - air temperature

Returns the tendency due to rain evaporation or snow sublimation.

ClimateMachine.Microphysics.snow_melt — Function

snow_melt(param_set, snow_param_set, q_sno, ρ)

param_set - abstract set with earth parameters

snow_param_set - abstract set with snow microphysics parameters

q_sno - snow water specific humidity

ρ - air density

Returns the tendency due to snow melt.