Temperature Profiles

ClimateMachine.TemperatureProfiles.TemperatureProfile — Type

TemperatureProfile

Specifies the temperature or virtual temperature profile for a reference state.

Instances of this type are required to be callable objects with the following signature

T,p = (::TemperatureProfile)(param_set::AbstractParameterSet, z::FT) where {FT}

where T is the temperature or virtual temperature (in K), and p is the pressure (in Pa).

ClimateMachine.TemperatureProfiles.IsothermalProfile — Function

IsothermalProfile(param_set, T_virt)
IsothermalProfile(param_set, ::Type{FT<:AbstractFloat})

A uniform virtual temperature profile, which is implemented as a special case of DecayingTemperatureProfile.

ClimateMachine.TemperatureProfiles.DryAdiabaticProfile — Type

DryAdiabaticProfile{FT} <: TemperatureProfile{FT}

A temperature profile that has uniform dry potential temperature θ

Fields

T_surface

Surface temperature (K)

T_min_ref

Minimum temperature (K)

ClimateMachine.TemperatureProfiles.DecayingTemperatureProfile — Type

DecayingTemperatureProfile{F} <: TemperatureProfile{FT}

A virtual temperature profile that decays smoothly with height z, from T_virt_surf to T_min_ref over a height scale H_t. The default height scale H_t is the density scale height evaluated with T_virt_surf.

\[T_{\text{v}}(z) = \max(T_{\text{v, sfc}} − (T_{\text{v, sfc}} - T_{\text{v, min}}) \tanh(z/H_{\text{t}})\]

Fields

T_virt_surf

Virtual temperature at surface (K)

T_min_ref

Minimum virtual temperature at the top of the atmosphere (K)

H_t

Height scale over which virtual temperature drops (m)