Spectra

Methods

ClimateMachine.Spectra.power_spectrum_1d — Function

power_spectrum_1d(::AtmosGCMConfigType, var_grid, z, lat, lon, weight)

Calculates the 1D (zonal) power spectra using the fourier transform at each latitude and level from a 3D velocity field. The snapshots of these spectra should be averaged to obtain a time-average. The input velocities must be interpolated to a Gaussian grid.

References

Aksel WIIN-NIELSEN (1967)

John N Koshyk , Kevin Hamilton (2001)

Arguments

var_grid: variable (typically u or v) on a Gausian (lon, lat, z) grid to be transformed

z: vertical coordinate (height or pressure)

lat: latitude

lon: longitude

mass_weight: for mass-weighted calculations

ClimateMachine.Spectra.power_spectrum_2d — Function

power_spectrum_2d(::AtmosGCMConfigType, var_grid, mass_weight)

transform variable on grid to the 2d spectral space using fft on latitude circles

(as for the 1D spectrum) and Legendre polynomials for meridians, and calculate spectra

Arguments

var_grid: variable (typically u or v) on a Gausian (lon, lat, z) grid to be transformed

mass_weight: weight for mass-weighted calculations

References

Ferdinand Baer (1972)

ClimateMachine.Spectra.power_spectrum_3d — Function

power_spectrum_3d(::AtmosLESConfigType, u, v, w, L, dim, nor)

Calculates the Powerspectrum from the 3D velocity fields u, v, w. Inputs need to be equi-spaced and the domain is assumed to be the same size and have the same number of points in all directions.

Arguments

L: size domain

dim: number of points

nor: normalization factor