Surface Fluxes

ClimateMachine.SurfaceFluxes — Module

SurfaceFluxes

Interface

surface_conditions computes

Monin-Obukhov length

Potential temperature flux (if not given) using Monin-Obukhov theory

transport fluxes using Monin-Obukhov theory

friction velocity/temperature scale/tracer scales

exchange coefficients

References

S Nishizawa , Y Kitamura (2018)

Universal Functions

ClimateMachine.SurfaceFluxes.UniversalFunctions — Module

UniversalFunctions

Universal stability and stability correction functions for SurfaceFluxes module.

ClimateMachine.SurfaceFluxes.UniversalFunctions.Gryanik — Type

Gryanik <: AbstractUniversalFunction{FT}

References

Vladimir M Gryanik , Christof L { \ " u } pkes , Andrey Grachev , Dmitry Sidorenko (2020)

Equations in reference:

ϕ_m: Eq. 13
ϕ_h: Eq. 13
ψ_m: Eq. 14
ψ_h: Eq. 14

Fields

param_set

Parameter set

L

Monin-Obhukov Length

ClimateMachine.SurfaceFluxes.UniversalFunctions.Grachev — Type

Grachev <: AbstractUniversalFunction{FT}

References

Andrey A Grachev , Edgar L Andreas , Christopher W Fairall , Peter S Guest , P Ola G Persson (2007)

Equations in reference:

ϕ_m: Eq. 13
ϕ_h: Eq. 13
ψ_m: Eq. 14
ψ_h: Eq. 14

Fields

param_set

Parameter set

L

Monin-Obhukov Length

ClimateMachine.SurfaceFluxes.UniversalFunctions.Businger — Type

Businger

Reference

S Nishizawa , Y Kitamura (2018)

Original research

Joost A Businger , John C Wyngaard , Yꎬ Izumi , Edward F Bradley (1971)

Equations in reference:

ϕ_m: Eq. A1
ϕ_h: Eq. A2
ψ_m: Eq. A3
ψ_h: Eq. A4

Fields

param_set

Parameter set

L

Monin-Obhukov Length

ClimateMachine.SurfaceFluxes.UniversalFunctions.phi — Function

phi

Universal stability function for wind shear (ϕm) and temperature gradient (ϕh)

ClimateMachine.SurfaceFluxes.UniversalFunctions.psi — Function

psi

Universal stability correction function for momentum (ψm) and heat (ψh)

ClimateMachine.SurfaceFluxes.UniversalFunctions.Psi — Function

Psi

Integral of universal stability correction function for momentum (ψm) and heat (ψh)

Fluxes API

ClimateMachine.SurfaceFluxes.surface_conditions — Function

surface_conditions

Surface conditions given

x_initial initial guess for solution (L_MO, u_star, θ_star, ϕ_star, ...)

x_ave volume-averaged value for variable x

x_s surface value for variable x

z_0 roughness length for variable x

F_exchange flux at the top for variable x

VDSE_scale virtual dry static energy (i.e., basic potential temperature)

Δz layer thickness (not spatial discretization)

z coordinate axis

a free model parameter with prescribed value of 4.7

VDSE_flux_star potential temperature flux (optional)

If VDSE_flux_star is not given, then it is computed by iteration of equations 3, 17, and 18 in Nishizawa2018.

ClimateMachine.SurfaceFluxes.exchange_coefficients — Function

exchange_coefficients(z, F_exchange, a, x_star, VDSE_scale, L_MO)

Computes exchange transfer coefficients

K_exchange exchange coefficients