TurbulenceClosures

ClimateMachine.TurbulenceClosures — Module

TurbulenceClosures

Functions for turbulence, sub-grid scale modelling. These include viscosity terms, diffusivity and stress tensors.

ConstantViscosity

ViscousSponge

SmagorinskyLilly

Vreman

AnisoMinDiss

Turbulence Closure Model Constructors

ClimateMachine.TurbulenceClosures.TurbulenceClosureModel — Type

Abstract type with default do-nothing behaviour for

arbitrary turbulence closure models.

ClimateMachine.TurbulenceClosures.WithDivergence — Type

WithDivergence

A divergence type which includes the divergence term in the momentum flux tensor

ClimateMachine.TurbulenceClosures.WithoutDivergence — Type

WithoutDivergence

A divergence type which does not include the divergence term in the momentum flux tensor

ClimateMachine.TurbulenceClosures.ConstantViscosity — Type

ConstantViscosity <: TurbulenceClosureModel

Abstract type for constant viscosity models

ClimateMachine.TurbulenceClosures.ConstantDynamicViscosity — Type

ConstantDynamicViscosity <: ConstantViscosity

Turbulence with constant dynamic viscosity (ρν). Divergence terms are included in the momentum flux tensor if divergence_type is WithDivergence.

Fields

ρν

Dynamic Viscosity [kg/m/s]

divergence_type

ClimateMachine.TurbulenceClosures.ConstantKinematicViscosity — Type

ConstantKinematicViscosity <: ConstantViscosity

Turbulence with constant kinematic viscosity (ν). Divergence terms are included in the momentum flux tensor if divergence_type is WithDivergence.

Fields

ν

Kinematic Viscosity [m2/s]

divergence_type

ClimateMachine.TurbulenceClosures.SmagorinskyLilly — Type

SmagorinskyLilly <: TurbulenceClosureModel

Fields

C_smag

Smagorinsky Coefficient [dimensionless]

Smagorinsky Model Reference

See Joseph Smagorinsky (1963)

Lilly Model Reference

See Douglas K Lilly (1962)

Brunt-Väisälä Frequency Reference

See Dale R Durran , Joseph B Klemp (1982)

ClimateMachine.TurbulenceClosures.Vreman — Type

Vreman{FT} <: TurbulenceClosureModel

Filter width Δ is the local grid resolution calculated from the mesh metric tensor. A Smagorinsky coefficient is specified and used to compute the equivalent Vreman coefficient.

ν_e = √(Bᵦ/(αᵢⱼαᵢⱼ)) where αᵢⱼ = ∂uⱼ∂uᵢ with uᵢ the resolved scale velocity component.

βij = Δ²αₘᵢαₘⱼ

Bᵦ = β₁₁β₂₂ + β₂₂β₃₃ + β₁₁β₃₃ - β₁₂² - β₁₃² - β₂₃²

βᵢⱼ is symmetric, positive-definite. If Δᵢ = Δ, then β = Δ²αᵀα

Fields

C_smag

Smagorinsky Coefficient [dimensionless]

Reference

AW Vreman (2004)

ClimateMachine.TurbulenceClosures.AnisoMinDiss — Type

AnisoMinDiss{FT} <: TurbulenceClosureModel

Filter width Δ is the local grid resolution calculated from the mesh metric tensor. A Poincare coefficient is specified and used to compute the equivalent AnisoMinDiss coefficient (computed as the solution to the eigenvalue problem for the Laplacian operator).

Fields

C_poincare

Reference

See Catherine A Vreugdenhil , John R Taylor (2018)

ClimateMachine.TurbulenceClosures.HyperDiffusion — Type

Abstract type for Hyperdiffusion models

ClimateMachine.TurbulenceClosures.NoHyperDiffusion — Type

NoHyperDiffusion <: HyperDiffusion Defines a default hyperdiffusion model with zero hyperdiffusive fluxes.

ClimateMachine.TurbulenceClosures.DryBiharmonic — Type

DryBiharmonic{FT} <: HyperDiffusion

Assumes dry compressible flow. Horizontal hyperdiffusion methods for application in GCM and LES settings Timescales are prescribed by the user while the diffusion coefficient is computed as a function of the grid lengthscale.

Fields

τ_timescale

ClimateMachine.TurbulenceClosures.EquilMoistBiharmonic — Type

EquilMoistBiharmonic{FT} <: HyperDiffusion

Assumes equilibrium thermodynamics in compressible flow. Horizontal hyperdiffusion methods for application in GCM and LES settings Timescales are prescribed by the user while the diffusion coefficient is computed as a function of the grid lengthscale.

Fields

τ_timescale

τ_timescale_q_tot

ClimateMachine.TurbulenceClosures.ViscousSponge — Type

Abstract type for viscous sponge layers.

Modifier for viscosity computed from existing turbulence closures.

ClimateMachine.TurbulenceClosures.NoViscousSponge — Type

NoViscousSponge

No modifiers applied to viscosity/diffusivity in sponge layer

Fields

ClimateMachine.TurbulenceClosures.UpperAtmosSponge — Type

Upper domain viscous relaxation

Applies modifier to viscosity and diffusivity terms in a user-specified upper domain sponge region

Fields

z_max

Maximum domain altitude (m)

z_sponge

Altitude at with sponge starts (m)

α_max

Sponge Strength 0 ⩽ α_max ⩽ 1

γ

Sponge exponent

Supporting Methods

ClimateMachine.TurbulenceClosures.turbulence_tensors — Function

ν, D_t, τ = turbulence_tensors(
                ::TurbulenceClosureModel,
                orientation::Orientation,
                param_set::AbstractParameterSet,
                state::Vars,
                diffusive::Vars,
                aux::Vars,
                t::Real
            )

Compute the kinematic viscosity (ν), the diffusivity (D_t) and SGS momentum flux tensor (τ) for a given turbulence closure. Each closure overloads this method with the appropriate calculations for the returned quantities.

Arguments

::TurbulenceClosureModel = Struct identifier for turbulence closure model

orientation = BalanceLaw.orientation

param_set parameter set

state = Array of prognostic (state) variables. See vars_state in BalanceLaw

diffusive = Array of diffusive variables

aux = Array of auxiliary variables

t = time

ClimateMachine.TurbulenceClosures.init_aux_turbulence! — Function

init_aux_turbulence!

Initialise auxiliary variables for turbulence models. Overload for specific turbulence closure type.

ClimateMachine.TurbulenceClosures.turbulence_nodal_update_auxiliary_state! — Function

turbulence_nodal_update_auxiliary_state!

Update auxiliary variables for turbulence models. Overload for specific turbulence closure type.

ClimateMachine.TurbulenceClosures.principal_invariants — Function

principal_invariants(X)

Calculates principal invariants of a tensor X. Returns 3 element tuple containing the invariants.

ClimateMachine.TurbulenceClosures.symmetrize — Function

symmetrize(X)

Given a (3,3) second rank tensor X, compute (X + X')/2, returning a symmetric SHermitianCompact object.

ClimateMachine.TurbulenceClosures.norm2 — Function

norm2(X)

Given a tensor X, computes X:X.

ClimateMachine.TurbulenceClosures.strain_rate_magnitude — Function

strain_rate_magnitude(S)

Given the rate-of-strain tensor S, computes its magnitude.