Coupling API

ClimateMachine.Coupling — Module

Coupling

Primitive coupling module sufficient for initial atmos-ocean-land coupled simulation.

Methods

ClimateMachine.Coupling.CplState — Type

CplState(;)

Type for holding Coupler "state". This is the namespace through which coupled components communicate. Its role is to provide a level of indirection so that components remain modular and so that any data communication, interpolation, reindexing/unit conversions and filtering etc... can be embeded in the intermdediate coupling layer.

To start with we can just use a dictionary key and value table that holds labelled pointers to various fields. A field is exported by one component and imported by one or more other components. Components can select which fields are needed by using the Dict symbols.

ClimateMachine.Coupling.CplTestModel — Type

CplTestModel(;domain,
              equations,
              nsteps,
              btags=((0,0),(0,0),(1,2) ),
              dt=1.,
              NFSecondOrder=CentralNumericalFluxSecondOrder())

Builds an instance of a coupler test model. This is a toy model used for testing and designing coupling machinery. In a full-blown coupled experiment this model would be replaced by a full compnent model.

  -  `domain` the spatial mesh used by this model. 
  -  `equations` the Balance Law used by this model.
  -  `nsteps` number of component steps to run during each coupling step.
  -  `btags` boundary labels for different edges.
  -  `dt` component timestep to use on each component step.
  -  `NFSecondOrder` numerical flux to use for second order terms.

Each returned model instance is independent and has its own grid, balance law, time stepper and other attributes. For now the code keeps some of these things the same for initial testing, including component timestepper and initial time (both of which need tweaking to use for real setups).

The argumetns and btags and NFSecondOrder are useful for this particular model. A real model might have many more flags and/or may wrap the component creation very differently. Any component should allow itself to set a number of timesteps to execute with a certain timestep to synchronize with the coupling time scale.

ClimateMachine.ODESolvers.CplSolver — Type

CplSolver( ;component_list, cdt, t0 = 0 )

A time stepping like object for advancing a coupled system made up of a pre-defined set of named components specified in component_list. Each component is a balance law, discretization and timestepper collection. The coupler will step them forward by a nsteps substeps to advance the coupled system by a simulated time cdt.

Components are registered with prestep() and poststep() functions. The prestep() functions get fields for use in the component from the coupler name space. The poststep() functions put fields for use by other components into te coupler name space. The CplSolver abstraction controls

the outer time stepping sequencing of components

the excution of actions mapping exports from one or more components to imports of other components through an intermediary coupler name space.

Some notes -

For now components need to include slightly wasteful "shadow" variables for accumulating boundary flux terms they compute across RK stages and across timesteps. These are defined within the component balance law. The shadown variable is a full 3d array because of the way the current infrastructure works. This can be tidied up later once design is settled.