Diagnostics

ClimateMachine.Diagnostics — Module

Diagnostics

Accumulate mean fields and covariance statistics on the computational grid.

Types

ClimateMachine.Diagnostics.DiagnosticsGroup — Type

DiagnosticsGroup

Holds a set of diagnostics that share a collection interval, a filename prefix, an output writer, an interpolation, and any extra parameters.

Diagnostics groups

A ClimateMachine driver may use any number of the methods described below to create DiagnosticsGroups which must be specified to the ClimateMachine in a DiagnosticsConfiguration.

ClimateMachine.Diagnostics.setup_atmos_default_diagnostics — Function

setup_atmos_default_diagnostics(
    ::AtmosLESConfigType,
    interval::String,
    out_prefix::String;
    writer = NetCDFWriter(),
    interpol = nothing,
)

Create and return a DiagnosticsGroup containing the "AtmosDefault" diagnostics for LES configurations. All the diagnostics in the group will run at the specified interval, be interpolated to the specified boundaries and resolution, and will be written to files prefixed by out_prefix using writer.

setup_atmos_default_diagnostics(
    ::AtmosGCMConfigType,
    interval::String,
    out_prefix::String;
    writer::AbstractWriter,
    interpol = nothing,
)

Create and return a DiagnosticsGroup containing the "AtmosDefault" diagnostics for GCM configurations. All the diagnostics in the group will run at the specified interval, be interpolated to the specified boundaries and resolution, and will be written to files prefixed by out_prefix using writer.

ClimateMachine.Diagnostics.setup_atmos_core_diagnostics — Function

setup_atmos_core_diagnostics(
    ::AtmosLESConfigType,
    interval::String,
    out_prefix::String;
    writer::AbstractWriter,
    interpol = nothing,
)

Create and return a DiagnosticsGroup containing the "AtmosLESCore" diagnostics for LES configurations. All the diagnostics in the group will run at the specified interval, be interpolated to the specified boundaries and resolution, and will be written to files prefixed by out_prefix using writer.

ClimateMachine.Diagnostics.setup_atmos_default_perturbations — Function

setup_atmos_default_perturbations(
    ::AtmosLESConfigType,
    interval::String,
    out_prefix::String;
    writer::AbstractWriter,
    interpol = nothing,
)

Create and return a DiagnosticsGroup containing the "AtmosLESDefaultPerturbations" diagnostics for the LES configuration. All the diagnostics in the group will run at the specified interval, and written to files prefixed by out_prefix using writer.

ClimateMachine.Diagnostics.setup_atmos_refstate_perturbations — Function

setup_atmos_refstate_perturbations(
    ::ClimateMachineConfigType,
    interval::String,
    out_prefix::String;
    writer = NetCDFWriter(),
    interpol = nothing,
)

Create and return a DiagnosticsGroup containing the "AtmosRefStatePerturbations" diagnostics for both LES and GCM configurations. All the diagnostics in the group will run at the specified interval, be interpolated to the specified boundaries and resolution, and written to files prefixed by out_prefix using writer.

ClimateMachine.Diagnostics.setup_atmos_turbulence_stats — Function

setup_atmos_turbulence_stats(
    ::ClimateMachineConfigType,
    interval::String,
    out_prefix::String,
    nor::Float64,
    iter::Float64;
    writer = NetCDFWriter(),
    interpol = nothing,
)

Create and return a DiagnosticsGroup containing the "AtmosTurbulenceStats" diagnostics for both LES and GCM configurations. All the diagnostics in the group will run at the specified interval, be interpolated to the specified boundaries and resolution, and written to files prefixed by out_prefix using writer.

ClimateMachine.Diagnostics.setup_atmos_mass_energy_loss — Function

setup_atmos_mass_energy_loss(
    ::ClimateMachineConfigType,
    interval::String,
    out_prefix::String,
    writer = NetCDFWriter(),
    interpol = nothing,
)

Create and return a DiagnosticsGroup containing the "AtmosMassEnergyLoss" diagnostics for Atmos LES and GCM configurations. All the diagnostics in the group will run at the specified interval, be interpolated to the specified boundaries and resolution, and written to files prefixed by out_prefix using writer.

ClimateMachine.Diagnostics.setup_atmos_spectra_diagnostics — Function

setup_atmos_spectra_diagnostics(
    ::AtmosLESConfigType,
    interval::String,
    out_prefix::String;
    writer = NetCDFWriter(),
    interpol = nothing,
    nor = Inf,
)

Create and return a DiagnosticsGroup containing a diagnostic for Atmos LES configurations that dumps the spectrum at the specified interval after the velocity fields have been interpolated, into NetCDF files prefixed by out_prefix.

setup_atmos_spectra_diagnostics(
    ::AtmosGCMConfigType,
    interval::String,
    out_prefix::String;
    writer = NetCDFWriter(),
    interpol = nothing,
    nor = 1.0,
)

Create and return a DiagnosticsGroup containing a diagnostic that dumps the spectrum at the specified interval after the velocity fields have been interpolated, into NetCDF files prefixed by out_prefix.

ClimateMachine.Diagnostics.setup_dump_state_diagnostics — Function

setup_dump_state_diagnostics(
    ::ClimateMachineConfigType,
    interval::String,
    out_prefix::String;
    writer = NetCDFWriter(),
    interpol = nothing,
)

Create and return a DiagnosticsGroup containing a diagnostic that simply dumps the conservative state variables at the specified interval after being interpolated, into NetCDF files prefixed by out_prefix.

ClimateMachine.Diagnostics.setup_dump_aux_diagnostics — Function

setup_dump_aux_diagnostics(
    ::ClimateMachineConfigType,
    interval::String,
    out_prefix::String;
    writer = NetCDFWriter(),
    interpol = nothing,
)

Create and return a DiagnosticsGroup containing a diagnostic that simply dumps the auxiliary state variables at the specified interval after being interpolated, into NetCDF files prefixed by out_prefix.