InputOutput

VTK

Functions

ClimateMachine.VTK.writevtk — Function

writevtk(prefix, Q::MPIStateArray, dg::DGModel [, fieldnames])

Write a vtk file for all the fields in the state array Q using geometry and connectivity information from dg.grid. The filename will start with prefix which may also contain a directory path. The names used for each of the fields in the vtk file can be specified through the collection of strings fieldnames; if not specified the fields names will be "Q1" through "Qk" where k is the number of states in Q, i.e., k = size(Q,2).

writevtk(prefix, Q::MPIStateArray, dg::DGModel, fieldnames,
         state_auxiliary::MPIStateArray, auxfieldnames)

Write a vtk file for all the fields in the state array Q and auxiliary state state_auxiliary using geometry and connectivity information from dg.grid. The filename will start with prefix which may also contain a directory path. The names used for each of the fields in the vtk file can be specified through the collection of strings fieldnames and auxfieldnames.

If fieldnames === nothing then the fields names will be "Q1" through "Qk" where k is the number of states in Q, i.e., k = size(Q,2).

If auxfieldnames === nothing then the fields names will be "aux1" through "auxk" where k is the number of states in state_auxiliary, i.e., k = size(state_auxiliary,2).

writegrid(prefix, grid::DiscontinuousSpectralElementGrid)

Write a vtk file for the grid. The filename will start with prefix which may also contain a directory path.

ClimateMachine.VTK.writevtk_helper — Function

writevtk_helper(prefix, vgeo::Array, Q::Array, grid, fieldnames)

Internal helper function for writevtk