Driver

ClimateMachine.parse_commandline — Function

parse_commandline(defaults::Union{Nothing, Dict{Symbol,Any}),
                  custom_settings::Union{Nothing,ArgParseSettings}=nothing)

Parse process command line ARGS values. If defaults is not nothing, override default values for cli argument defaults. If custom_settings arg is not nothing, add ArgParseSettings to the parsing step.

Returns a Dict containing parsed process ARGS values.

ClimateMachine.array_type — Function

ClimateMachine.array_type()

Return the array type used by ClimateMachine. This defaults to (CPU-based) Array and is only correctly set (based on choice from the command line, from an environment variable, or from experiment code) after ClimateMachine.init() is called.

ClimateMachine.init — Function

ClimateMachine.init(;init_driver::Bool=true, kwargs...)

Perform necessary initializations for ClimateMachine:

Initialize MPI.

Parse command line arguments. To support experiment-specific arguments,

arg_settings may be specified (it is an ArgParse.ArgParseSettings); it will be imported into ClimateMachine's settings.

Determine whether GPU(s) is available and should be used (pass

disable-gpu = true if not) and set the ClimateMachine array type appropriately.

Set up the global logger.

Setting init_driverd = false will set the ClimateMachine.Settings singleton values without initializing the ClimateMachine driver runtime.

ClimateMachine.Settings values can be overloaded at runtime upon initialization. If keyword argument overloads are not supplied, the init routine will try and fallback on any CLIMATEMACHINE_SETTINGS_<VALUE> ENV variables defined for the process, otherwise the defaulting to ClimateMachine.Settings.

Keyword Arguments

disable_gpu::Bool = false: do not use the GPU

show_updates::String = "60secs": interval at which to show simulation updates

diagnostics::String = "never": interval at which to collect diagnostics"

vtk::String = "never": inteverval at which to write simulation vtk output

monitor_timestep_duration::String = "never": interval in time-steps at which to output wall-clock time per time-step

monitor_courant_numbers::String = "never": interval at which to output acoustic, advective, and diffusive Courant numbers"

checkpoint::String = "never": interval at which to output a checkpoint

checkpoint_keep_one::Bool = true: (interval) keep all checkpoints (instead of just the most recent)"

checkpoint_at_end::Bool = false: create a checkpoint at the end of the simulation"

checkpoint_dir::String = "checkpoint": absolute or relative path to checkpoint directory

restart_from_num::Int = -1: checkpoint number from which to restart (in checkpoint_dir)

fix_rng_seed::Bool = false: set RNG seed to a fixed value for reproducibility

log_level::String = "INFO": log level for ClimateMachine global default runtime logger

disable_custom_logger::String = false: disable using a global custom logger for ClimateMachine

output_dir::String = "output": (path) absolute or relative path to output data directory

integration_testing::Bool = false: enable integration_testing

ClimateMachine.invoke! — Function

ClimateMachine.invoke!(
    solver_config::SolverConfiguration;
    diagnostics_config = nothing,
    user_callbacks = (),
    check_euclidean_distance = false,
    adjustfinalstep = false,
    user_info_callback = (init) -> nothing,
)

Run the simulation defined by solver_config.

Keyword Arguments:

The user_callbacks are passed to the ODE solver as callback functions; see [`ODESolvers.solve!]@ref().

If check_euclidean_distance is true, then the Euclidean distance between the final solution and initial condition function evaluated withsolver_config.timeend` is reported.

The value of 'adjustfinalstepis passed to the ODE solver; see [ODESolvers.solve!]@ref().

The function user_info_callback is called after the default info callback (which is called every Settings.show_updates interval). The single input argument init is true when the callback is called for initialization (before time stepping begins) and false when called during the actual ODE solve; see GenericCallbacks and [`ODESolvers.solve!]@ref().

ClimateMachine.DriverConfiguration — Type

ClimateMachine.DriverConfiguration

Collects all parameters necessary to set up a ClimateMachine simulation.

ClimateMachine.SolverConfiguration — Type

ClimateMachine.SolverConfiguration

Parameters needed by ClimateMachine.solve!() to run a simulation.