Input/Output

Artifact Wrappers

ClimateMachine.ArtifactWrappers.ArtifactFile — Type

ArtifactFile

A single data file to be downloaded, containing both the url and the name to use locally.

Fields

url

URL pointing to data to be downloaded

filename

Local name used for downloaded online data

ClimateMachine.ArtifactWrappers.get_data_folder — Function

get_data_folder(art_wrap::ArtifactWrapper)

Get local folder of dataset defined in art_wrap.

Example:

dataset_path = get_data_folder(dataset)

ClimateMachine.ArtifactWrappers.ArtifactWrapper — Type

ArtifactWrapper

A set of data files to be downloaded, grouped by data_name. Example:

dataset = ArtifactWrapper(
    @__DIR__,
    isempty(get(ENV, "CI", "")),
    "MyDataSet",
    ArtifactFile[
        ArtifactFile(
            url="https://..../SomeNetCDF1.nc",
            filename="experiment1.nc",
        ),
        ArtifactFile(
            url="https://..../SomeNetCDF2.nc",
            filename="experiment2.nc",
        ),
    ]
)

Fields

artifact_dir

Directory to store artifact / data

local_run

Locally running (not using CI)

artifact_toml

Path to the used Artifacts.toml

data_name

Unique name of dataset

artifact_files

Array of ArtifactFile's, grouped by this dataset

VTK

ClimateMachine.VTK.writevtk — Function

writevtk(prefix, Q::MPIStateArray, dg::SpaceDiscretization [, fieldnames];
         number_sample_points = 0)

Write a vtk file for all the fields in the state array Q using geometry and connectivity information from dg.grid. The filename will start with prefix which may also contain a directory path. The names used for each of the fields in the vtk file can be specified through the collection of strings fieldnames; if not specified the fields names will be "Q1" through "Qk" where k is the number of states in Q, i.e., k = size(Q,2).

If number_sample_points > 0 then the fields are sampled on an equally spaced, tensor-product grid of points with 'numbersamplepoints' in each direction and the output VTK element type is set to by a VTK lagrange type.

When number_sample_points == 0 the raw nodal values are saved, and linear VTK elements are used connecting the degree of freedom boxes.

writevtk(prefix, Q::MPIStateArray, dg::SpaceDiscretization, fieldnames,
         state_auxiliary::MPIStateArray, auxfieldnames;
         number_sample_points = 0)

Write a vtk file for all the fields in the state array Q and auxiliary state state_auxiliary using geometry and connectivity information from dg.grid. The filename will start with prefix which may also contain a directory path. The names used for each of the fields in the vtk file can be specified through the collection of strings fieldnames and auxfieldnames.

If fieldnames === nothing then the fields names will be "Q1" through "Qk" where k is the number of states in Q, i.e., k = size(Q,2).

If auxfieldnames === nothing then the fields names will be "aux1" through "auxk" where k is the number of states in state_auxiliary, i.e., k = size(state_auxiliary,2).

If number_sample_points > 0 then the fields are sampled on an equally spaced, tensor-product grid of points with 'numbersamplepoints' in each direction and the output VTK element type is set to by a VTK lagrange type.

When number_sample_points == 0 the raw nodal values are saved, and linear VTK elements are used connecting the degree of freedom boxes.

ClimateMachine.VTK.writepvtu — Function

writepvtu(pvtuprefix, vtkprefixes, fieldnames, FT)

Write a pvtu file with the prefix 'pvtuprefix' for the collection of vtk files given by 'vtkprefixes' using names of fields 'fieldnames'. The data in the vtu files is of type FT.

Writers

ClimateMachine.Writers.init_data — Function

init_data(
    writer,
    filename,
    dims,
    vars,
)

Creates the specified file, initializing it with the specified dimension information. An unlimited time dimension is implicitly created. The specified variables are also defined. This function must be called before append_data(). Specialized by every Writer subtype.

Arguments:

- writer: instance of a subtype of AbstractWriter.

- filename: into which to write data (without extension).

- dims: Dict of dimension name to 2-tuple of dimension values and Dict

of attributes.

- vars: Dict of variable name to 3-tuple of a k-tuple of dimension

names, variable type, and Dict of attributes.

ClimateMachine.Writers.full_name — Function

full_name(
    writer,
    filename,
)

Appends the appropriate (based on writer) extension to the specified filename.

ClimateMachine.Writers.append_data — Function

append_data(
    writer,
    filename,
    varvals,
    simtime,
)

Appends the specified variables to the specified file. The file must have been previously created with init_data(). simtime is appended to the time dimension variable. Specialized by every Writer subtype.

Arguments:

- writer: instance of a subtype of AbstractWriter.

- filename: into which to write data (without extension).

- varvals: Dict of variable name to k-dimensional array of values.

- simtime: Current simulation time.