Runoff

Runoff types and functions

ClimateMachine.Land.Runoff.AbstractPrecipModel — Type

AbstractPrecipModel{FT <: AbstractFloat}

ClimateMachine.Land.Runoff.DrivenConstantPrecip — Type

DrivenConstantPrecip{FT, F} <: AbstractPrecipModel{FT}

Instance of a precipitation distribution where the precipication value is constant across the domain. However, this value can change in time.

Fields

mp

Mean precipitation in grid

ClimateMachine.Land.Runoff.AbstractSurfaceRunoffModel — Type

AbstractSurfaceRunoffModel

Abstract type for different surface runoff models. Currently, only NoRunoff is supported.

ClimateMachine.Land.Runoff.NoRunoff — Type

NoRunoff <: AbstractSurfaceRunoffModel

Chosen when no runoff is to be modeled.

ClimateMachine.Land.Runoff.compute_surface_flux — Function

compute_surface_flux(
    runoff_model::AbstractSurfaceRunoffModel,
    precip_model::AbstractPrecipModel{FT},
    state::Vars,
    t::Real,
) where {FT}

Given a runoff model and a precipitation distribution function, compute the surface water flux. This can be a function of time, and state.