Soil Heat Parameterizations

Heat functions

ClimateMachine.Land.SoilHeatParameterizations.volumetric_heat_capacity — Function

volumetric_heat_capacity(
    θ_l::FT,
    θ_i::FT,
    ρc_ds::FT,
    param_set::AbstractParameterSet
) where {FT}

Compute the expression for volumetric heat capacity.

ClimateMachine.Land.SoilHeatParameterizations.volumetric_internal_energy — Function

volumetric_internal_energy(
    θ_i::FT,
    ρc_s::FT,
    T::FT,
    param_set::AbstractParameterSet
) where {FT}

Compute the expression for volumetric internal energy.

ClimateMachine.Land.SoilHeatParameterizations.saturated_thermal_conductivity — Function

saturated_thermal_conductivity(
    θ_l::FT,
    θ_i::FT,
    κ_sat_unfrozen::FT,
    κ_sat_frozen::FT
) where {FT}

Compute the expression for saturated thermal conductivity of soil matrix.

ClimateMachine.Land.SoilHeatParameterizations.thermal_conductivity — Function

thermal_conductivity(
    κ_dry::FT,
    K_e::FT,
    κ_sat::FT
) where {FT}

Compute the expression for thermal conductivity of soil matrix.

ClimateMachine.Land.SoilHeatParameterizations.relative_saturation — Function

relative_saturation(
        θ_l::FT,
        θ_i::FT,
        porosity::FT
) where {FT}

Compute the expression for relative saturation.

ClimateMachine.Land.SoilHeatParameterizations.kersten_number — Function

kersten_number(
    θ_i::FT,
    S_r::FT,
    soil_param_functions::PS
) where {FT, PS}

Compute the expression for the Kersten number.

ClimateMachine.Land.SoilHeatParameterizations.k_solid — Function

function k_solid(
    ν_ss_om::FT,
    ν_ss_quartz::FT,
    κ_quartz::FT,
    κ_minerals::FT,
    κ_om::FT,
) where {FT}

Computes the thermal conductivity of the solid material in soil. The _ss_ subscript denotes that the volumetric fractions of the soil components are referred to the soil solid components, not including the pore space.

ClimateMachine.Land.SoilHeatParameterizations.k_dry — Function

function k_dry(
    param_set::AbstractParameterSet
    soil_param_functions::PS,
) where {PS}

Computes the thermal conductivity of dry soil.

ClimateMachine.Land.SoilHeatParameterizations.ksat_unfrozen — Function

function ksat_unfrozen(
    κ_solid::FT,
    porosity::FT,
    κ_l::FT
) where {FT}

Computes the thermal conductivity for saturated unfrozen soil.

ClimateMachine.Land.SoilHeatParameterizations.ksat_frozen — Function

function ksat_frozen(
    κ_solid::FT,
    porosity::FT,
    κ_ice::FT
) where {FT}

Computes the thermal conductivity for saturated frozen soil.

ClimateMachine.Land.SoilHeatParameterizations.volumetric_internal_energy_liq — Function

volumetric_internal_energy_liq(
    T::FT,
    T_ref::FT,
) where {FT}

Compute the expression for the volumetric internal energy of liquid water.

ClimateMachine.Land.SoilHeatParameterizations.temperature_from_ρe_int — Function

function temperature_from_ρe_int(
    ρe_int::FT,
    θ_i::FT,
    ρc_s::FT,
    param_set::AbstractParameterSet
) where {FT}

Computes the temperature of soil given θ_i and volumetric internal energy ρe_int.