Numerical Fluxes

Types

ClimateMachine.DGMethods.FVReconstructions.AbstractReconstruction — Type

AbstractReconstruction

Supertype for FV reconstructions.

Concrete types must provide implementions of - width(recon) returns the width of the reconstruction. Total number of points used in reconstruction of top and bottom states is 2width(recon) + 1 - (::AbstractReconstruction)(statetop, statebottom, cellstates::SVector, cellweights) compute the reconstruction

(::AbstractReconstruction)(
        state_top,
        state_bottom,
        cell_states,
        cell_weights,
    )

Perform the finite volume reconstruction for the top and bottom states using the tuple of cell_states values using the cell_weights.

ClimateMachine.DGMethods.FVReconstructions.AbstractSlopeLimiter — Type

AbstractSlopeLimiter

Supertype for FV slope limiter

Given two values Δ1 and Δ2, should return the value of

Δ2 * ϕ(Δ1 / Δ2)

where 0 ≤ ϕ(r) ≤ 2 is the slope limiter

ClimateMachine.DGMethods.FVReconstructions.width — Function

width(recon::AbstractReconstruction)

Returns the width of the stencil need for the FV reconstruction recon. Total number of values used in the reconstruction are 2width(recon) + 1

ClimateMachine.DGMethods.FVReconstructions.FVConstant — Type

FVConstant <: AbstractReconstruction

Reconstruction type for cell centered finite volume methods (e.g., constants)

ClimateMachine.DGMethods.FVReconstructions.FVLinear — Type

FVLinear{W = 1} <: AbstractReconstruction

Reconstruction type for limited linear reconstruction finite volume methods.

Note

The optional type parameter W is mainly for debugging purposes and allows the stencil to be artificially widened to make sure the kernels work with wide stencils.

FVLinear(limiter = VanLeer())
FVLinear{W}(limiter = VanLeer())

Construct the FVLinear reconstruction type with the given slope limiter

ClimateMachine.DGMethods.FVReconstructions.VanLeer — Type

VanLeer <: AbstractSlopeLimiter

Classic Van-Leer limiter

   ϕ(r) = (r + |r|) / (1 + r)

References

@article{van1974towards,
  title = {Towards the ultimate conservative difference scheme. II.
           Monotonicity and conservation combined in a second-order scheme},
  author = {Van Leer, Bram},
  journal = {Journal of computational physics},
  volume = {14},
  number = {4},
  pages = {361--370},
  year = {1974},
  doi = {10.1016/0021-9991(74)90019-9}
}