Arrays

MPIStateArrays

Storage for the state of a discretization.

CLIMA.MPIStateArrays.MPIStateArray — Type

MPIStateArray{FT, DATN<:AbstractArray{FT,3}, DAI1, DAV,
              DAT2<:AbstractArray{FT,2}} <: AbstractArray{FT, 3}

CLIMA.MPIStateArrays.begin_ghost_exchange! — Function

begin_ghost_exchange!(Q::MPIStateArray; dependencies = nothing)

Begin the MPI halo exchange of the data stored in Q. A KernelAbstractions Event is returned that can be used as a dependency to end the exchange.

CLIMA.MPIStateArrays.end_ghost_exchange! — Function

end_ghost_exchange!(Q::MPIStateArray; dependencies = nothing)

This function blocks on the host until the ghost halo is received from MPI. A KernelAbstractions Event is returned that can be waited on to indicate when the data is ready on the device.