Diagnostics

Diagnostic Fields

Diagnostics computes the physically relevant fields for postprocessing.

Types

CLIMA.Diagnostics.VecGrad — Type

VecGrad{FT <: AbstractFloat,
     FTA2D <: AbstractArray{FT, 2},
     FTA3D <: AbstractArray{FT, 3},}

This data structure stores the spatial gradients of a velocity field.

Fields

data

Device array storing the spatial gradient data

∂₁u₁

View of ∂u₁/∂x₁

∂₂u₁

View of ∂u₁/∂x₂

∂₃u₁

View of ∂u₁/∂x₃

∂₁u₂

View of ∂u₂/∂x₁

∂₂u₂

View of ∂u₂/∂x₂

∂₃u₂

View of ∂u₂/∂x₃

∂₁u₃

View of ∂u₃/∂x₁

∂₂u₃

View of ∂u₃/∂x₂

∂₃u₃

View of ∂u₃/∂x₃

Usage

VecGrad(Npl, Nel, ::Type{FT}) where {FT <: AbstractFloat}

Arguments for the inner constructor

Npl: Number of local degrees of freedom in a spectral element

Nel: Number of spectral elements

FT: Floating point precision

CLIMA.Diagnostics.Vorticity — Type

Vorticity{FT <: AbstractFloat,
       FTA2D <: AbstractArray{FT, 2},
       FTA3D <: AbstractArray{FT, 3},}

This data structure stores the vorticity of a velocity field.

Fields

data

Device array storing the vorticity data

Ω₁

View of x1 component of vorticity

Ω₂

View of x2 component of vorticity

Ω₃

View of x3 component of vorticity

Usage

Vorticity(Npl, Nel, ::Type{FT}) where {FT <: AbstractFloat}

Arguments for the inner constructor

Npl: Number of local degrees of freedom in a spectral element

Nel: Number of spectral elements

FT: Floating point precision

Functions

CLIMA.Diagnostics.compute_vec_grad — Function

compute_vec_grad(model::BalanceLaw,
                     Q::AbstractArray{FT},
                    dg::DGModel,
                 vgrad::VecGrad{FT},) where {FT <: AbstractFloat}

This function computes the spatial gradients of the velocity field.

Arguments

model: BalanceLaw

Q: State array

dg: DGmodel

vgrad: Vector gradients

CLIMA.Diagnostics.compute_vorticity — Function

compute_vorticity(dg::DGModel,
               vgrad::VecGrad{FT},
                vort::Vorticity{FT},) where {FT <: AbstractFloat}

This function computes the vorticity of the velocity field.

Arguments

dg: DGmodel

vgrad: Velocity gradients

vort: Vorticity