LinearSolvers

GeneralizedConjugateResidual

CLIMA.GeneralizedConjugateResidualSolver.GeneralizedConjugateResidual — Type

Conjugate Residual

GeneralizedConjugateResidual(K, Q; rtol, atol)

This is an object for solving linear systems using an iterative Krylov method. The constructor parameter K is the number of steps after which the algorithm is restarted (if it has not converged), Q is a reference state used only to allocate the solver internal state, and tolerance specifies the convergence criterion based on the relative residual norm. The amount of memory required by the solver state is roughly (2K + 2) * size(Q). This object is intended to be passed to the linearsolve! command.

This uses the restarted Generalized Conjugate Residual method of Eisenstat (1983).

References

@article{eisenstat1983variational,
  title={Variational iterative methods for nonsymmetric systems of linear equations},
  author={Eisenstat, Stanley C and Elman, Howard C and Schultz, Martin H},
  journal={SIAM Journal on Numerical Analysis},
  volume={20},
  number={2},
  pages={345--357},
  year={1983},
  publisher={SIAM}
}

GeneralizedMinimalResidual

CLIMA.GeneralizedMinimalResidualSolver.GeneralizedMinimalResidual — Type

GMRES

GeneralizedMinimalResidual(Q; M, rtol, atol)

This is an object for solving linear systems using an iterative Krylov method. The constructor parameter M is the number of steps after which the algorithm is restarted (if it has not converged), Q is a reference state used only to allocate the solver internal state, and rtol specifies the convergence criterion based on the relative residual norm. The amount of memory required for the solver state is roughly (M + 1) * size(Q). This object is intended to be passed to the linearsolve! command.

This uses the restarted Generalized Minimal Residual method of Saad and Schultz (1986).

References

@article{saad1986gmres,
  title={GMRES: A generalized minimal residual algorithm for solving nonsymmetric linear systems},
  author={Saad, Youcef and Schultz, Martin H},
  journal={SIAM Journal on scientific and statistical computing},
  volume={7},
  number={3},
  pages={856--869},
  year={1986},
  publisher={SIAM}
}

ConjugateGradientSolver

CLIMA.ConjugateGradientSolver.ConjugateGradient — Type

ConjugateGradient

function ConjugateGradient(Q::AT; rtol = eps(eltype(Q)), atol = eps(eltype(Q)), dims = :) where AT

Description

Outer constructor for the ConjugateGradient struct

Arguments

Q:(array). The kind of object that linearoperator! acts on.

Keyword Arguments

rtol: (float). relative tolerance

atol: (float). absolute tolerance

dims: (tuple or : ). the dimensions to reduce over

reshape_tuple: (tuple). the dimensions that the conjugate gradient solver operators over

Comment

The reshape tuple is necessary in case the linearoperator! is defined over vectors of a different size as compared to what plays nicely with the dimension reduction in the ConjugateGradient. It also allows the user to define preconditioners over arrays that are more convenienently shaped.

Return

ConjugateGradient struct

function ConjugateGradient(Q::MPIStateArray; rtol = eps(eltype(Q)), atol = eps(eltype(Q)), dims = :)

Description

Outer constructor for the ConjugateGradient struct with MPIStateArrays. THIS IS A HACK DUE TO RESHAPE FUNCTIONALITY ON MPISTATEARRAYS.

Arguments

Q:(array). The kind of object that linearoperator! acts on.

Keyword Arguments

rtol: (float). relative tolerance

atol: (float). absolute tolerance

dims: (tuple or : ). the dimensions to reduce over

reshape_tuple: (tuple). the dimensions that the conjugate gradient solver operators over

Comment

The reshape tuple is necessary in case the linearoperator! is defined over vectors of a different size as compared to what plays nicely with the dimension reduction in the ConjugateGradient. It also allows the user to define preconditioners over arrays that are more convenienently shaped.

Return

ConjugateGradient struct

CLIMA.LinearSolvers.initialize! — Function

LS.initialize!(linearoperator!, Q, Qrhs, solver::ColumnwisePreconditionedConjugateGradient, args...)

Description

This function initializes the iterative solver. It is called as part of the AbstractIterativeLinearSolver routine. SEE CODEREF for documentation on AbstractIterativeLinearSolver

Arguments

linearoperator!: (function). This applies the predefined linear operator on an array. Applies a linear operator to object "y" and overwrites object "z". The function argument i s linearoperator!(z,y, args...) and it returns nothing.

Q: (array). This is an object that linearoperator! outputs

Qrhs: (array). This is an object that linearoperator! acts on

solver: (struct). This is a scruct for dispatch, in this case for ColumnwisePreconditionedConjugateGradient

args...: (arbitrary). This is optional arguments that can be passed into linearoperator! function.

Keyword Arguments

There are no keyword arguments

Return

converged: (bool). A boolean to say whether or not the iterative solver has converged.

threshold: (float). The value of the residual for the first timestep

Comment

This function does nothing for conjugate gradient

CLIMA.LinearSolvers.doiteration! — Function

LS.doiteration!(linearoperator!, Q, Qrhs, solver::ColumnwisePreconditionedConjugateGradient, threshold, args...; applyPC!)

Description

This function enacts the iterative solver. It is called as part of the AbstractIterativeLinearSolver routine. SEE CODEREF for documentation on AbstractIterativeLinearSolver

Arguments

linearoperator!: (function). This applies the predefined linear operator on an array. Applies a linear operator to object "y" and overwrites object "z". It is a function with arguments linearoperator!(z,y, args...), where "z" gets overwritten by "y" and "args..." are additional arguments passed to the linear operator. The linear operator is assumed to return nothing.

Q: (array). This is an object that linearoperator! overwrites

Qrhs: (array). This is an object that linearoperator! acts on. This is the rhs to the linear system

solver: (struct). This is a scruct for dispatch, in this case for ConjugateGradient

threshold: (float). Either an absolute or relative tolerance

applyPC!: (function). Applies a preconditioner to objecy "y" and overwrites object "z". applyPC!(z,y)

args...: (arbitrary). This is necessary for the linearoperator! function which has a signature linearoperator!(b, x, args....)

Keyword Arguments

There are no keyword arguments

Return

converged: (bool). A boolean to say whether or not the iterative solver has converged.

iteration: (int). Iteration number for the iterative solver

threshold: (float). The value of the residual for the first timestep

Comment

This function does conjugate gradient

LS.doiteration!(linearoperator!, Q::MPIStateArray, Qrhs::MPIStateArray, solver::ColumnwisePreconditionedConjugateGradient, threshold, args...; applyPC!)

Description

This function enacts the iterative solver. It is called as part of the AbstractIterativeLinearSolver routine. SEE CODEREF for documentation on AbstractIterativeLinearSolver. THIS IS A HACK TO WORK WITH MPISTATEARRAYS. THE ISSUE IS WITH RESHAPE.

Arguments

linearoperator!: (function). This applies the predefined linear operator on an array. Applies a linear operator to object "y" and overwrites object "z". It is a function with arguments linearoperator!(z,y, args...), where "z" gets overwritten by "y" and "args..." are additional arguments passed to the linear operator. The linear operator is assumed to return nothing.

Q: (array). This is an object that linearoperator! overwrites

Qrhs: (array). This is an object that linearoperator! acts on. This is the rhs to the linear system

solver: (struct). This is a scruct for dispatch, in this case for ConjugateGradient

threshold: (float). Either an absolute or relative tolerance

applyPC!: (function). Applies a preconditioner to objecy "y" and overwrites object "z". applyPC!(z,y)

args...: (arbitrary). This is necessary for the linearoperator! function which has a signature linearoperator!(b, x, args....)

Keyword Arguments

There are no keyword arguments

Return

converged: (bool). A boolean to say whether or not the iterative solver has converged.

iteration: (int). Iteration number for the iterative solver

threshold: (float). The value of the residual for the first timestep

Comment

This function does conjugate gradient

LinearSolvers

CLIMA.LinearSolvers.AbstractLinearSolver — Type

AbstractLinearSolver

This is an abstract type representing a generic linear solver.

CLIMA.LinearSolvers.AbstractIterativeLinearSolver — Type

AbstractIterativeLinearSolver

This is an abstract type representing a generic iterative linear solver.

The available concrete implementations are:

GeneralizedConjugateResidual

GeneralizedMinimalResidual

CLIMA.LinearSolvers.linearsolve! — Function

linearsolve!(linearoperator!, solver::AbstractIterativeLinearSolver, Q, Qrhs, args...)

Solves a linear problem defined by the linearoperator! function and the state Qrhs, i.e,

\[L(Q) = Q_{rhs}\]

using the solver and the initial guess Q. After the call Q contains the solution. The arguments args is passed to linearoperator! when it is called.

CLIMA.LinearSolvers.settolerance! — Function

settolerance!(solver::AbstractIterativeLinearSolver, tolerance, relative)

Sets the relative or absolute tolerance of the iterative linear solver solver to tolerance.

CLIMA.LinearSolvers.prefactorize — Function

prefactorize(linop!, linearsolver, args...)

Prefactorize the in-place linear operator linop! for use with linearsolver.