Balance Laws

The balance law

ClimateMachine.BalanceLaws.BalanceLaw — Type

BalanceLaw

An abstract type representing a PDE balance law of the form:

\[q_{,t} + Σ_{i=1,...d} F_{i,i} = s\]

Subtypes L should define the methods below.

State variable types

ClimateMachine.BalanceLaws.Prognostic — Type

Prognostic <: AbstractStateType

Prognostic variables in the PDE system, which are specified by the BalanceLaw, and solved for by the ODE solver.

ClimateMachine.BalanceLaws.Auxiliary — Type

Auxiliary <: AbstractStateType

Auxiliary variables help serve several purposes:

Pre-compute and store "expensive" variables, for example, quantities computed in vertical integrals.

Diagnostic exports

ClimateMachine.BalanceLaws.Gradient — Type

Gradient <: AbstractStateType

Variables whose gradients must be computed.

ClimateMachine.BalanceLaws.GradientFlux — Type

GradientFlux <: AbstractStateType

Flux variables, which are functions of gradients.

ClimateMachine.BalanceLaws.GradientLaplacian — Type

GradientLaplacian <: AbstractStateType

Gradient-Laplacian variables.

ClimateMachine.BalanceLaws.Hyperdiffusive — Type

Hyperdiffusive <: AbstractStateType

Hyper-diffusive variables

ClimateMachine.BalanceLaws.UpwardIntegrals — Type

UpwardIntegrals <: AbstractStateType

Variables computed in upward integrals

ClimateMachine.BalanceLaws.DownwardIntegrals — Type

DownwardIntegrals <: AbstractStateType

Variables computed in downward integrals

Variable specification methods

ClimateMachine.BalanceLaws.vars_state — Function

vars_state(::L, ::AbstractStateType, FT)

A tuple of symbols containing the state variables given a float type FT.

Initial condition methods

ClimateMachine.BalanceLaws.init_state_prognostic! — Function

init_state_prognostic!(
    ::L,
    state_prognostic::Vars,
    state_auxiliary::Vars,
    coords,
    args...,
)

Initialize the prognostic state variables at $t = 0$.

init_state_prognostic!(::HBModel)

sets the initial value for state variables dispatches to oceaninitstate! which is defined in a problem file such as SimpleBoxProblem.jl

init_state_prognostic!(
    m::AtmosModel,
    state::Vars,
    aux::Vars,
    coords,
    t,
    args...,
)

Initialise state variables. args... provides an option to include configuration data (current use cases include problem constants, spline-interpolants).

ClimateMachine.BalanceLaws.init_state_auxiliary! — Function

init_state_auxiliary!(
    ::L,
    state_auxiliary::Vars,
    geom::LocalGeometry,
)

Initialize the auxiliary state, at $t = 0$.

init_state_auxiliary!(::HBModel)

sets the initial value for auxiliary variables (those that aren't related to vertical integrals) dispatches to oceaninitaux! which is defined in a problem file such as SimpleBoxProblem.jl

No need to init, initialize by full model

init_state_auxiliary!(
    m::AtmosModel,
    aux::Vars,
    geom::LocalGeometry,
)

Initialise auxiliary variables for each AtmosModel subcomponent. Store Cartesian coordinate information in aux.coord.

ClimateMachine.BalanceLaws.nodal_init_state_auxiliary! — Function

nodal_init_state_auxiliary!()

Nodal initialization the auxiliary state variables at $t = 0$

Source term kernels

ClimateMachine.BalanceLaws.flux_first_order! — Function

flux_first_order!(
    ::L,
    flux::Grad,
    state_prognostic::Vars,
    state_auxiliary::Vars,
    t::Real,
    directions
)

Compute first-order flux terms in balance law equation.

function flux_first_order!(
    rem_balance_law::RemBL,
    flux::Grad,
    state::Vars,
    aux::Vars,
    t::Real,
    directions,
)

Evaluate the remainder flux by first evaluating the main flux and subtracting the subcomponent fluxes.

Only models which have directions that are included in the directions tuple are evaluated. When these models are evaluated the models underlying direction is passed (not the original directions argument).

flux_first_order!(::HBModel)

calculates the hyperbolic flux contribution to state variables this computation is done pointwise at each nodal point

arguments:

m -> model in this case HBModel F -> array of fluxes for each state variable Q -> array of state variables A -> array of aux variables t -> time, not used

computations

∂ᵗu = ∇⋅(g*η + g∫αᵀθdz + v⋅u) ∂ᵗθ = ∇⋅(vθ) where v = (u,v,w)

flux_first_order!(
    m::AtmosModel,
    flux::Grad,
    state::Vars,
    aux::Vars,
    t::Real
)

Computes and assembles non-diffusive fluxes in the model equations.

ClimateMachine.BalanceLaws.flux_second_order! — Function

flux_second_order!(
    ::L,
    flux::Grad,
    state_prognostic::Vars,
    state_gradient_flux::Vars,
    hyperdiffusive::Vars,
    state_auxiliary::Vars,
    t::Real
)

Compute second-order flux terms in balance law equation.

flux_second_order!(::HBModel)

calculates the parabolic flux contribution to state variables this computation is done pointwise at each nodal point

arguments:

m: model in this case HBModel

F: array of fluxes for each state variable

Q: array of state variables

D: array of diff variables

A: array of aux variables

t: time, not used

computations

∂ᵗu = -∇⋅(ν∇u) ∂ᵗθ = -∇⋅(κ∇θ)

flux_second_order!(::HBModel)

calculates the parabolic flux contribution to state variables this computation is done pointwise at each nodal point

arguments:

m: model in this case HBModel

F: array of fluxes for each state variable

Q: array of state variables

D: array of diff variables

A: array of aux variables

t: time, not used

computations

∂ᵗu = -∇⋅(ν∇u) ∂ᵗθ = -∇⋅(κ∇θ)

flux_second_order!(
    atmos::AtmosModel,
    flux::Grad,
    state::Vars,
    diffusive::Vars,
    hyperdiffusive::Vars,
    aux::Vars,
    t::Real
)

Diffusive fluxes in AtmosModel. Viscosity, diffusivity are calculated in the turbulence subcomponent and accessed within the diffusive flux function. Contributions from subcomponents are then assembled (pointwise).

ClimateMachine.BalanceLaws.source! — Function

source!(
    ::L,
    source::Vars,
    state_prognostic::Vars,
    diffusive::Vars,
    state_auxiliary::Vars,
    t::Real
)

Compute non-conservative source terms in balance law equation.

function source!(
    rem_balance_law::RemBL,
    source::Vars,
    state::Vars,
    diffusive::Vars,
    aux::Vars,
    t::Real,
    directions,
)

Evaluate the remainder source by first evaluating the main source and subtracting the subcomponent sources.

Only models which have directions that are included in the directions tuple are evaluated. When these models are evaluated the models underlying direction is passed (not the original directions argument).

source!(::HBModel)
calculates the source term contribution to state variables
this computation is done pointwise at each nodal point

arguments:
m -> model in this case HBModel
F -> array of fluxes for each state variable
Q -> array of state variables
A -> array of aux variables
t -> time, not used

computations
∂ᵗu = -f×u
∂ᵗη = w|(z=0)

source!(
    m::AtmosModel,
    source::Vars,
    state::Vars,
    diffusive::Vars,
    aux::Vars,
    t::Real,
    direction::Direction,
)

Computes (and assembles) source terms S(Y) in:

∂Y
-- = - ∇ • F + S(Y)
∂t

Integral kernels

ClimateMachine.BalanceLaws.indefinite_stack_integral! — Function

indefinite_stack_integral!

Compute indefinite integral along stack.

ClimateMachine.BalanceLaws.reverse_indefinite_stack_integral! — Function

reverse_indefinite_stack_integral!

Compute reverse indefinite integral along stack.

ClimateMachine.BalanceLaws.integral_load_auxiliary_state! — Function

integral_load_auxiliary_state!

Specify how to compute integrands. Can be functions of the prognostic state and auxiliary variables.

ClimateMachine.BalanceLaws.integral_set_auxiliary_state! — Function

integral_set_auxiliary_state!

Specify which auxiliary variables are used to store the output of the integrals.

ClimateMachine.BalanceLaws.reverse_integral_load_auxiliary_state! — Function

reverse_integral_load_auxiliary_state!

Specify auxiliary variables need their integrals reversed.

ClimateMachine.BalanceLaws.reverse_integral_set_auxiliary_state! — Function

reverse_integral_set_auxiliary_state!

Specify which auxiliary variables are used to store the output of the reversed integrals.

Gradient/Laplacian kernels

ClimateMachine.BalanceLaws.compute_gradient_flux! — Function

compute_gradient_flux!(
    ::L,
    state_gradient_flux::Vars,
    ∇transformstate::Grad,
    state_auxiliary::Vars,
    t::Real
)

Transformation of gradients to the diffusive variables.

compute_gradient_flux!(::HBModel)

copy ∇u and ∇θ to var_diffusive this computation is done pointwise at each nodal point

arguments:

m: model in this case HBModel

D: array of diffusive variables

G: array of gradient variables

Q: array of state variables

A: array of aux variables

t: time, not used

compute_gradient_flux!(::LinearHBModel)

copy ν∇u and κ∇θ to var_diffusive this computation is done pointwise at each nodal point

arguments:

m: model in this case HBModel

D: array of diffusive variables

G: array of gradient variables

Q: array of state variables

A: array of aux variables

t: time, not used

compute_gradient_flux!(::TurbulenceClosureModel, _...)

Post-gradient-transformed variables specific to turbulence models.

ClimateMachine.BalanceLaws.compute_gradient_argument! — Function

compute_gradient_argument!(
    ::L,
    transformstate::Vars,
    state_prognostic::Vars,
    state_auxiliary::Vars,
    t::Real
)

Transformation of state variables to variables of which gradients are computed.

compute_gradient_argument!(::HBModel)

copy u and θ to var_gradient this computation is done pointwise at each nodal point

arguments:

m: model in this case HBModel

G: array of gradient variables

Q: array of state variables

A: array of aux variables

t: time, not used

compute_gradient_argument!(::LinearHBModel)

copy u and θ to var_gradient this computation is done pointwise at each nodal point

arguments:

m: model in this case HBModel

G: array of gradient variables

Q: array of state variables

A: array of aux variables

t: time, not used

compute_gradient_argument!

Assign pre-gradient-transform variables specific to turbulence models.

ClimateMachine.BalanceLaws.transform_post_gradient_laplacian! — Function

transform_post_gradient_laplacian!(
    ::L,
    Qhypervisc_div::Vars,
    ∇Δtransformstate::Grad,
    state_auxiliary::Vars,
    t::Real
)

Transformation of Laplacian gradients to the hyperdiffusive variables.

Auxiliary kernels

ClimateMachine.BalanceLaws.wavespeed — Function

wavespeed(
    ::L,
    n⁻,
    state_prognostic::Vars,
    state_auxiliary::Vars,
    t::Real,
    direction
)

Wavespeed.

ClimateMachine.BalanceLaws.boundary_state! — Function

boundary_state!(
    ::NumericalFluxGradient,
    ::L,
    state_prognostic⁺::Vars,
    state_auxiliary⁺::Vars,
    normal⁻,
    state_prognostic⁻::Vars,
    state_auxiliary⁻::Vars,
    bctype,
    t
)
boundary_state!(
    ::NumericalFluxFirstOrder,
    ::L,
    state_prognostic⁺::Vars,
    state_auxiliary⁺::Vars,
    normal⁻,
    state_prognostic⁻::Vars,
    state_auxiliary⁻::Vars,
    bctype,
    t
)
boundary_state!(
    ::NumericalFluxSecondOrder,
    ::L,
    state_prognostic⁺::Vars,
    state_gradient_flux⁺::Vars,
    state_auxiliary⁺:
    Vars, normal⁻,
    state_prognostic⁻::Vars,
    state_gradient_flux⁻::Vars,
    state_auxiliary⁻::Vars,
    bctype,
    t
)

Apply boundary conditions for

NumericalFluxGradient numerical flux (internal method)

NumericalFluxFirstOrder first-order unknowns

NumericalFluxSecondOrder second-order unknowns

ClimateMachine.BalanceLaws.update_auxiliary_state! — Function

update_auxiliary_state!(
    dg::DGModel,
    m::BalanceLaw,
    Q::MPIStateArray,
    t::Real,
    elems::UnitRange,
)

Update the auxiliary state variables with global scope.

ClimateMachine.BalanceLaws.update_auxiliary_state_gradient! — Function

update_auxiliary_state_gradient!

Update the auxiliary state gradient variables.

ClimateMachine.BalanceLaws.nodal_update_auxiliary_state! — Function

nodal_update_auxiliary_state!()

Update the auxiliary state variables at each location in space.