Balance Laws

The balance law

ClimateMachine.BalanceLaws.BalanceLaw — Type

abstract type BalanceLaw end

An abstract type representing a PDE balance law of the form:

\[\frac{dq}{dt} = \nabla \cdot F_1(q, a, t) + \nabla \cdot F_2(q, \nabla g, h, a, t) + S(q, \nabla g, a, t)\]

where:

$q$ is the prognostic state,

$a$ is the auxiliary state,

$g = G(q, a, t)$ is the gradient state (variables of which we compute the gradient),

$h$ is the hyperdiffusive state.

Subtypes of BalanceLaw should define the following interfaces:

vars_state to define the prognostic, auxiliary and intermediate variables.

flux_first_order! to compute $F_1$

flux_second_order! to compute $F_2$

source! to compute $S$

If vars(bl, ::GradientFlux, FT) is non-empty, then the following should be defined:

compute_gradient_argument! to compute $G$

compute_gradient_flux! is a linear transformation of $\nabla g$

If vars(bl, ::Hyperdiffusive, FT) is non-empty, then the following should be defined:

transform_post_gradient_laplacian!

Additional functions:

wavespeed if using the Rusanov numerical flux.

boundary_state! if using non-periodic boundary conditions.

Tendency types and methods

ClimateMachine.BalanceLaws.PrognosticVariable — Type

PrognosticVariable

Subtypes are used for specifying each prognostic variable.

ClimateMachine.BalanceLaws.AbstractOrder — Type

AbstractOrder

Subtypes are used for dispatching on the flux order.

ClimateMachine.BalanceLaws.AbstractTendencyType — Type

AbstractTendencyType

Subtypes are used for specifying a tuple of tendencies to be accumulated.

ClimateMachine.BalanceLaws.TendencyDef — Type

TendencyDef

Subtypes are used for specifying each tendency definition.

ClimateMachine.BalanceLaws.eq_tends — Function

eq_tends(::PrognosticVariable, ::BalanceLaw, ::AbstractTendencyType)

A tuple of TendencyDefs given

PrognosticVariable prognostic variable

AbstractTendencyType tendency type

BalanceLaw balance law

i.e., a tuple of TendencyDefs corresponding to F₁, F₂, or S for a single prognostic variable in:

`∂_t Yᵢ + (∇•F₁(Y))ᵢ + (∇•F₂(Y,G)))ᵢ = (S(Y,G))ᵢ`

ClimateMachine.BalanceLaws.prognostic_vars — Function

prognostic_vars(::BalanceLaw)

A tuple of PrognosticVariables given the BalanceLaw.

i.e., a tuple of PrognosticVariables corresponding to the column-vector Yᵢ in:

`∂_t Yᵢ + (∇•F₁(Y))ᵢ + (∇•F₂(Y,G)))ᵢ = (S(Y,G))ᵢ`

ClimateMachine.BalanceLaws.precompute — Function

precompute(bl, args, ::AbstractTendencyType)

A nested NamedTuple of precomputed (cached) values and or objects. This is useful for "expensive" point-wise quantities that are used in multiple tendency terms. For example, computing a quantity that requires iteration.

ClimateMachine.BalanceLaws.fluxes — Function

fluxes(bl::BalanceLaw, order::O) where {O <: AbstractOrder}

A tuple of TendencyDef{Flux{O}}s given the BalanceLaw and the order::O.

i.e., a tuple of TendencyDef{Flux{O}}s corresponding to the column-vector F₁ or F₂ given the flux order order::O in:

`∂_t Yᵢ + (∇•F₁(Y))ᵢ + (∇•F₂(Y,G)))ᵢ = (S(Y,G))ᵢ`

ClimateMachine.BalanceLaws.sources — Function

sources(bl::BalanceLaw)

A tuple of TendencyDef{Source}s given the BalanceLaw.

i.e., a tuple of TendencyDef{Source}s corresponding to the column-vector S in:

`∂_t Yᵢ + (∇•F₁(Y))ᵢ + (∇•F₂(Y,G)))ᵢ = (S(Y,G))ᵢ`

ClimateMachine.BalanceLaws.show_tendencies — Function

show_tendencies(bl; include_params = false)

Show a table of the tendencies for each prognostic variable for the given balance law.

Using include_params = true will include the type parameters for each of the tendency fluxes and sources.

Requires definitions for

prognostic_vars

eq_tends

for the balance law.

Methods for fluxes and sources

ClimateMachine.BalanceLaws.flux — Function

flux

An individual flux. See BalanceLaw for more info.

ClimateMachine.BalanceLaws.source — Function

source

An individual source. See BalanceLaw for more info.

ClimateMachine.BalanceLaws.Σfluxes — Function

Σfluxes(fluxes::NTuple, args...)

Sum of the fluxes where

fluxes is an NTuple{N, TendencyDef{Flux{O}, PV}} where {N, PV, O}

ClimateMachine.BalanceLaws.Σsources — Function

Σsources(sources::NTuple, args...)

Sum of the sources where

sources is an NTuple{N, TendencyDef{Source, PV}} where {N, PV}

State variable types

ClimateMachine.BalanceLaws.AbstractStateType — Type

AbstractStateType

Subtypes of this describe the variables used by different parts of a BalanceLaw:

Prognostic

Primitive

Auxiliary

Gradient

GradientFlux

GradientLaplacian

Hyperdiffusive

UpwardIntegrals

DownwardIntegrals

See also vars_state.

ClimateMachine.BalanceLaws.Prognostic — Type

Prognostic <: AbstractStateType

Prognostic variables in the PDE system, which are specified by the BalanceLaw, and solved for by the ODE solver.

ClimateMachine.BalanceLaws.Primitive — Type

Primitive <: AbstractStateType

Primitive variables, which are specified by the BalanceLaw.

ClimateMachine.BalanceLaws.Auxiliary — Type

Auxiliary <: AbstractStateType

Auxiliary variables help serve several purposes:

Pre-compute and store "expensive" variables, for example, quantities computed in vertical integrals.

Diagnostic exports

ClimateMachine.BalanceLaws.Gradient — Type

Gradient <: AbstractStateType

Variables whose gradients must be computed.

ClimateMachine.BalanceLaws.GradientFlux — Type

GradientFlux <: AbstractStateType

Flux variables, which are functions of gradients.

ClimateMachine.BalanceLaws.GradientLaplacian — Type

GradientLaplacian <: AbstractStateType

Gradient-Laplacian variables.

ClimateMachine.BalanceLaws.Hyperdiffusive — Type

Hyperdiffusive <: AbstractStateType

Hyper-diffusive variables

ClimateMachine.BalanceLaws.UpwardIntegrals — Type

UpwardIntegrals <: AbstractStateType

Variables computed in upward integrals

ClimateMachine.BalanceLaws.DownwardIntegrals — Type

DownwardIntegrals <: AbstractStateType

Variables computed in downward integrals

Variable specification methods

ClimateMachine.BalanceLaws.vars_state — Function

BalanceLaws.vars_state(::BL, ::AbstractStateType, FT)

Defines the state variables of a BalanceLaw subtype BL with floating point type FT.

For each AbstractStateType, this should return a NamedTuple type, with element type either FT, an SArray with element type FT or another NamedTuple satisfying the same property.

For convenience, we recommend using the VariableTemplates.@vars macro.

Example

struct MyBalanceLaw <: BalanceLaw end

BalanceLaws.vars_state(::MyBalanceLaw, ::Prognostic, FT) =
    @vars(x::FT, y::SVector{3, FT})
BalanceLaws.vars_state(::MyBalanceLaw, ::Auxiliary, FT) =
    @vars(components::@vars(a::FT, b::FT))

Initial condition methods

ClimateMachine.BalanceLaws.init_state_prognostic! — Function

init_state_prognostic!(
    ::BL,
    state_prognostic::Vars,
    state_auxiliary::Vars,
    localgeo,
    args...,
)

Sets the initial state of the prognostic variables state_prognostic at each node for a BalanceLaw subtype BL.

ClimateMachine.BalanceLaws.init_state_auxiliary! — Function

init_state_auxiliary!(
    ::BL,
    statearray_auxiliary,
    geom::LocalGeometry,
)

Sets the initial state of the auxiliary variables state_auxiliary at each node for a BalanceLaw subtype BL. By default this calls nodal_init_state_auxiliary!.

init_state_auxiliary!(
    bl::BalanceLaw,
    f!,
    statearray_auxiliary,
    grid,
    direction;
    state_temporary = nothing
)

Apply f!(bl, state_auxiliary, tmp, geom) at each node, storing the result in statearray_auxiliary, where tmp are the values at the corresponding node in state_temporary and geom contains the geometry information.

init_state_auxiliary!(::HBModel)

sets the initial value for auxiliary variables (those that aren't related to vertical integrals) dispatches to oceaninitaux! which is defined in a problem file such as SimpleBoxProblem.jl

No need to init, initialize by full model

init_state_auxiliary!(
    m::AtmosModel,
    aux::Vars,
    grid,
    direction
)

Initialise auxiliary variables for each AtmosModel subcomponent. Store Cartesian coordinate information in aux.coord.

ClimateMachine.BalanceLaws.nodal_init_state_auxiliary! — Function

nodal_init_state_auxiliary!(::BL, state_auxiliary, state_temporary, geom)

Sets the initial state of the auxiliary variables state_auxiliary at each node for a BalanceLaw subtype BL.

See also init_state_auxiliary!.

Source term kernels

ClimateMachine.BalanceLaws.flux_first_order! — Function

flux_first_order!(
    ::BL,
    flux::Grad,
    state_prognostic::Vars,
    state_auxiliary::Vars,
    t::Real,
    direction
)

Sets the first-order (hyperbolic) flux terms for a BalanceLaw subtype BL.

ClimateMachine.BalanceLaws.flux_second_order! — Function

flux_second_order!(
    ::BL,
    flux::Grad,
    state_prognostic::Vars,
    state_gradient_flux::Vars,
    hyperdiffusive::Vars,
    state_auxiliary::Vars,
    t::Real
)

Sets second-order (parabolic) flux terms for a BalanceLaw subtype BL.

ClimateMachine.BalanceLaws.source! — Function

source!(
    ::BL,
    source::Vars,
    state_prognostic::Vars,
    diffusive::Vars,
    state_auxiliary::Vars,
    t::Real
)

Compute non-conservative source terms for a BalanceLaw subtype BL.

Integral kernels

ClimateMachine.BalanceLaws.indefinite_stack_integral! — Function

indefinite_stack_integral!

Compute indefinite integral along stack.

ClimateMachine.BalanceLaws.reverse_indefinite_stack_integral! — Function

reverse_indefinite_stack_integral!

Compute reverse indefinite integral along stack.

ClimateMachine.BalanceLaws.integral_load_auxiliary_state! — Function

integral_load_auxiliary_state!(::BL, integrand, state_prognostic, state_aux)

Specify variables integrand which will have their upward integrals computed.

See also UpwardIntegrals

ClimateMachine.BalanceLaws.integral_set_auxiliary_state! — Function

integral_set_auxiliary_state!(::BL, state_aux, integral)

Update auxiliary variables based on the upward integral integral defined in integral_load_auxiliary_state!.

ClimateMachine.BalanceLaws.reverse_integral_load_auxiliary_state! — Function

reverse_integral_load_auxiliary_state!(::BL, integrand, state_prognostic, state_aux)

Specify variables integrand which will have their downward integrals computed.

See also DownwardIntegrals

ClimateMachine.BalanceLaws.reverse_integral_set_auxiliary_state! — Function

reverse_integral_set_auxiliary_state!(::BL, state_aux, integral)

Update auxiliary variables based on the downward integral integral defined in reverse_integral_load_auxiliary_state!.

Gradient/Laplacian kernels

ClimateMachine.BalanceLaws.compute_gradient_flux! — Function

compute_gradient_flux!(
    ::BL,
    state_gradient_flux::Vars,
    ∇transformstate::Grad,
    state_prognostic::Vars,
    state_auxiliary::Vars,
    t::Real
)

Transformation of gradients to the diffusive variables for a BalanceLaw subtype BL. This should be a linear function of ∇transformstate

ClimateMachine.BalanceLaws.compute_gradient_argument! — Function

compute_gradient_argument!(
    ::BL,
    transformstate::Vars,
    state_prognostic::Vars,
    state_auxiliary::Vars,
    t::Real
)

Transformation of state variables state_prognostic to variables transformstate of which gradients are computed for a BalanceLaw subtype BL.

ClimateMachine.BalanceLaws.transform_post_gradient_laplacian! — Function

transform_post_gradient_laplacian!(
    ::BL,
    Qhypervisc_div::Vars,
    ∇Δtransformstate::Grad,
    state_auxiliary::Vars,
    t::Real
)

Transformation of Laplacian gradients to the hyperdiffusive variables for a BalanceLaw subtype BL.

Boundary conditions

ClimateMachine.BalanceLaws.boundary_conditions — Function

boundary_conditions(::BL)

Define the set of boundary conditions for the balance law BL. This should return a tuple, where a boundary tagged with the integer i will use the ith element of the tuple.

ClimateMachine.BalanceLaws.boundary_state! — Function

boundary_state!(
    ::NumericalFluxGradient,
    ::BC
    ::BL,
    state_prognostic⁺::Vars,
    state_auxiliary⁺::Vars,
    normal⁻,
    state_prognostic⁻::Vars,
    state_auxiliary⁻::Vars,
    t
)
boundary_state!(
    ::NumericalFluxFirstOrder,
    ::BC
    ::BL,
    state_prognostic⁺::Vars,
    state_auxiliary⁺::Vars,
    normal⁻,
    state_prognostic⁻::Vars,
    state_auxiliary⁻::Vars,
    t
)
boundary_state!(
    ::NumericalFluxSecondOrder,
    ::BC
    ::BL,
    state_prognostic⁺::Vars,
    state_gradient_flux⁺::Vars,
    state_auxiliary⁺:
    Vars, normal⁻,
    state_prognostic⁻::Vars,
    state_gradient_flux⁻::Vars,
    state_auxiliary⁻::Vars,
    t
)

Specify the opposite (+ side) face for the boundary condition type BC with balance law BL.

NumericalFluxGradient numerical flux (internal method)

NumericalFluxFirstOrder first-order unknowns

NumericalFluxSecondOrder second-order unknowns

Auxiliary kernels

ClimateMachine.BalanceLaws.wavespeed — Function

wavespeed(
    ::BL,
    n⁻,
    state_prognostic::Vars,
    state_auxiliary::Vars,
    t::Real,
    direction
)

Wavespeed in the direction n⁻ for a BalanceLaw subtype BL. This is required to be defined if using a RusanovNumericalFlux numerical flux.

ClimateMachine.BalanceLaws.update_auxiliary_state! — Function

update_auxiliary_state!(
    dg::DGModel,
    m::BalanceLaw,
    statearray_aux,
    t::Real,
    elems::UnitRange,
    [diffusive=false]
)

Hook to update the auxiliary state variables before calling any other functions.

By default, this calls nodal_update_auxiliary_state! at each node.

If diffusive=true, then state_gradflux is also passed to nodal_update_auxiliary_state!.

update_auxiliary_state!(::HBModel)

Applies the vertical filter to the zonal and meridional velocities to preserve numerical incompressibility Applies an exponential filter to θ to anti-alias the non-linear advective term

Doesn't actually touch the aux variables any more, but we need a better filter interface than this anyways

ClimateMachine.BalanceLaws.update_auxiliary_state_gradient! — Function

update_auxiliary_state_gradient!(
    dg::DGModel,
    m::BalanceLaw,
    statearray_aux,
    t::Real,
    elems::UnitRange,
    [diffusive=false]
)

Hook to update the auxiliary state variables after the gradient computation.

By default, this calls nothing.

If diffusive=true, then state_gradflux is also passed to nodal_update_auxiliary_state!.

ClimateMachine.BalanceLaws.nodal_update_auxiliary_state! — Function

nodal_update_auxiliary_state!(::BL, state_prognostic, state_auxiliary, [state_gradflux,] t)

Update the auxiliary state variables at each node for a BalanceLaw subtype BL. By default it does nothing.

Called by update_auxiliary_state!.