Thermodynamics

ClimateMachine.Thermodynamics — Module

Thermodynamics

Moist thermodynamic functions, e.g., for air pressure (atmosphere equation of state), latent heats of phase transitions, saturation vapor pressures, and saturation specific humidities.

AbstractParameterSet's

Many functions defined in this module rely on CLIMAParameters.jl. CLIMAParameters.jl defines several functions (e.g., many planet parameters). For example, to compute the mole-mass ratio:

using CLIMAParameters.Planet: molmass_ratio
using CLIMAParameters: AbstractEarthParameterSet
struct EarthParameterSet <: AbstractEarthParameterSet end
param_set = EarthParameterSet()
_molmass_ratio = molmass_ratio(param_set)

Because these parameters are widely used throughout this module, param_set is an argument for many Thermodynamics functions.

Numerical methods for saturation adjustment

Saturation adjustment function are designed to accept

sat_adjust_method a type used to dispatch which numerical method to use

and a function to return an instance of the numerical method. For example:

sa_numerical_method_ρpq returns an instance of the numerical method. One of these functions must be defined for the particular numerical method and the particular formulation (ρ-p-q_tot in this case).

The currently supported numerical methods, in RootSolvers.jl, are:

NewtonsMethod uses Newton method with analytic gradients

NewtonsMethodAD uses Newton method with autodiff

SecantMethod uses Secant method

RegulaFalsiMethod uses Regula-Falsi method

Thermodynamic State Constructors

ClimateMachine.Thermodynamics.PhasePartition — Type

PhasePartition

Represents the mass fractions of the moist air mixture.

Constructors

PhasePartition(q_tot::Real[, q_liq::Real[, q_ice::Real]])
PhasePartition(ts::ThermodynamicState)

See also PhasePartition_equil

Fields

tot

total specific humidity

liq

liquid water specific humidity (default: 0)

ice

ice specific humidity (default: 0)

ClimateMachine.Thermodynamics.PhasePartition_equil — Function

PhasePartition_equil(param_set, T, ρ, q_tot, phase_type)

Partition the phases in equilibrium, returning a PhasePartition object using the liquid_fraction function where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

q_tot total specific humidity

phase_type a thermodynamic state type

The residual q.tot - q.liq - q.ice is the vapor specific humidity.

ClimateMachine.Thermodynamics.ThermodynamicState — Type

ThermodynamicState{FT}

A thermodynamic state, which can be initialized for various thermodynamic formulations (via its sub-types). All ThermodynamicState's have access to functions to compute all other thermodynamic properties.

ClimateMachine.Thermodynamics.PhaseDry — Type

PhaseDry{FT} <: ThermodynamicState

A dry thermodynamic state (q_tot = 0).

Constructors

PhaseDry(param_set, e_int, ρ)

Fields

param_set

parameter set, used to dispatch planet parameter function calls

e_int

internal energy

ρ

density of dry air

ClimateMachine.Thermodynamics.PhaseDry_pT — Function

PhaseDry_pT(param_set, p, T)

Constructs a PhaseDry thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

p pressure

T temperature

ClimateMachine.Thermodynamics.PhaseDry_pθ — Function

PhaseDry_pθ(param_set, p, θ_dry)

Constructs a PhaseDry thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

p pressure

θ_dry dry potential temperature

ClimateMachine.Thermodynamics.PhaseDry_ρθ — Function

PhaseDry_ρθ(param_set, ρ, θ_dry)

Constructs a PhaseDry thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

ρ density

θ_dry dry potential temperature

ClimateMachine.Thermodynamics.PhaseDry_ρT — Function

PhaseDry_ρT(param_set, ρ, T)

Constructs a PhaseDry thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

ρ density

T temperature

ClimateMachine.Thermodynamics.PhaseDry_ρp — Function

PhaseDry_ρp(param_set, ρ, p)

Constructs a PhaseDry thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

ρ density

p pressure

ClimateMachine.Thermodynamics.PhaseEquil — Type

PhaseEquil{FT} <: ThermodynamicState

A thermodynamic state assuming thermodynamic equilibrium (therefore, saturation adjustment may be needed).

Constructors

PhaseEquil(param_set, e_int, ρ, q_tot)

Fields

param_set

parameter set, used to dispatch planet parameter function calls

e_int

internal energy

ρ

density of air (potentially moist)

q_tot

total specific humidity

T

temperature: computed via saturation_adjustment

ClimateMachine.Thermodynamics.PhaseEquil_ρTq — Function

PhaseEquil_ρTq(param_set, ρ, T, q_tot)

Constructs a PhaseEquil thermodynamic state from temperature.

param_set an AbstractParameterSet, see the Thermodynamics for more details

ρ density

T temperature

q_tot total specific humidity

ClimateMachine.Thermodynamics.PhaseEquil_pTq — Function

PhaseEquil_pTq(param_set, p, T, q_tot)

Constructs a PhaseEquil thermodynamic state from temperature.

param_set an AbstractParameterSet, see the Thermodynamics for more details

p pressure

T temperature

q_tot total specific humidity

ClimateMachine.Thermodynamics.PhaseEquil_pθq — Function

PhaseEquil_pθq(param_set, p, θ_liq_ice, q_tot)

Constructs a PhaseEquil thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

p pressure

θ_liq_ice liquid-ice potential temperature

q_tot total specific humidity

temperature_tol temperature tolerance for saturation adjustment

maxiter maximum iterations for saturation adjustment

ClimateMachine.Thermodynamics.PhaseEquil_ρθq — Function

PhaseEquil_ρθq(param_set, ρ, θ_liq_ice, q_tot)

Constructs a PhaseEquil thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

ρ (moist-)air density

θ_liq_ice liquid-ice potential temperature

q_tot total specific humidity

temperature_tol temperature tolerance for saturation adjustment

maxiter maximum iterations for saturation adjustment

ClimateMachine.Thermodynamics.PhaseEquil_ρpq — Function

PhaseEquil_ρpq(param_set, ρ, p, q_tot, perform_sat_adjust=true)

Constructs a PhaseEquil thermodynamic state from temperature.

param_set an AbstractParameterSet, see the Thermodynamics for more details

ρ density

p pressure

q_tot total specific humidity

perform_sat_adjust Bool indicating to perform saturation adjustment

maxiter maximum number of iterations to perform in saturation adjustment

sat_adjust_method the numerical method to use. See the Thermodynamics for options.

ClimateMachine.Thermodynamics.PhaseNonEquil — Type

PhaseNonEquil{FT} <: ThermodynamicState

A thermodynamic state assuming thermodynamic non-equilibrium (therefore, temperature can be computed directly).

Constructors

PhaseNonEquil(param_set, e_int, q::PhasePartition, ρ)

Fields

param_set

parameter set, used to dispatch planet parameter function calls

e_int

internal energy

ρ

density of air (potentially moist)

q

phase partition

ClimateMachine.Thermodynamics.PhaseNonEquil_ρTq — Function

PhaseNonEquil_ρTq(param_set, ρ, T, q_pt)

Constructs a PhaseNonEquil thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

ρ (moist-)air density

T temperature

q_pt phase partition

ClimateMachine.Thermodynamics.PhaseNonEquil_ρθq — Function

PhaseNonEquil_ρθq(param_set, ρ, θ_liq_ice, q_pt)

Constructs a PhaseNonEquil thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

ρ (moist-)air density

θ_liq_ice liquid-ice potential temperature

q_pt phase partition

and, optionally

potential_temperature_tol potential temperature for non-linear equation solve

maxiter maximum iterations for non-linear equation solve

ClimateMachine.Thermodynamics.PhaseNonEquil_pθq — Function

PhaseNonEquil_pθq(param_set, p, θ_liq_ice, q_pt)

Constructs a PhaseNonEquil thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

p pressure

θ_liq_ice liquid-ice potential temperature

q_pt phase partition

ClimateMachine.Thermodynamics.PhaseNonEquil_ρpq — Function

PhaseNonEquil_ρpq(param_set, ρ, p, q_pt)

Constructs a PhaseNonEquil thermodynamic state from:

param_set an AbstractParameterSet, see the Thermodynamics for more details

ρ density

p pressure

q_pt phase partition

Thermodynamic state methods

ClimateMachine.Thermodynamics.air_density — Function

air_density(param_set, T, p[, q::PhasePartition])

The (moist-)air density from the equation of state (ideal gas law) where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T air temperature

p pressure

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

air_density(ts::ThermodynamicState)

The (moist-)air density, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.air_pressure — Function

air_pressure(param_set, T, ρ[, q::PhasePartition])

The air pressure from the equation of state (ideal gas law) where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T air temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

air_pressure(ts::ThermodynamicState)

The air pressure from the equation of state (ideal gas law), given a thermodynamic state ts.

air_pressure(param_set, T::FT, T∞::FT, p∞::FT, ::DryAdiabaticProcess)

The air pressure for an isentropic process, where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

T∞ ambient temperature

p∞ ambient pressure

ClimateMachine.Thermodynamics.air_temperature — Function

air_temperature(param_set, e_int, q::PhasePartition)

The air temperature, where

param_set an AbstractParameterSet, see the Thermodynamics for more details

e_int internal energy per unit mass

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

air_temperature(ts::ThermodynamicState)

The air temperature, given a thermodynamic state ts.

air_temperature(param_set, p::FT, θ::FT, Φ::FT, ::DryAdiabaticProcess)

The air temperature for an isentropic process, where

param_set an AbstractParameterSet, see the Thermodynamics for more details

p pressure

θ potential temperature

ClimateMachine.Thermodynamics.condensate — Function

condensate(q::PhasePartition{FT})
condensate(ts::ThermodynamicState)

Condensate of the phase partition.

ClimateMachine.Thermodynamics.cp_m — Function

cp_m(param_set, [q::PhasePartition])

The isobaric specific heat capacity of moist air given

param_set an AbstractParameterSet, see the Thermodynamics for more details

and, optionally

q PhasePartition. Without this argument, the results are for dry air.

cp_m(ts::ThermodynamicState)

The isobaric specific heat capacity of moist air, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.cv_m — Function

cv_m(param_set, [q::PhasePartition])

The isochoric specific heat capacity of moist air given

param_set an AbstractParameterSet, see the Thermodynamics for more details

and, optionally

q PhasePartition. Without this argument, the results are for dry air.

cv_m(ts::ThermodynamicState)

The isochoric specific heat capacity of moist air, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.dry_pottemp — Function

dry_pottemp(param_set, T, ρ[, q::PhasePartition])

The dry potential temperature where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

dry_pottemp(ts::ThermodynamicState)

The dry potential temperature, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.exner — Function

exner(param_set, T, ρ[, q::PhasePartition)])

The Exner function where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

exner(ts::ThermodynamicState)

The Exner function, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.gas_constant_air — Function

gas_constant_air(param_set, [q::PhasePartition])

The specific gas constant of moist air given

param_set an AbstractParameterSet, see the Thermodynamics for more details

q PhasePartition. Without this argument, the results are for dry air.

gas_constant_air(ts::ThermodynamicState)

The specific gas constant of moist air given a thermodynamic state ts.

ClimateMachine.Thermodynamics.gas_constants — Function

(R_m, cp_m, cv_m, γ_m) = gas_constants(param_set, [q::PhasePartition])

Wrapper to compute all gas constants at once, where optionally,

param_set an AbstractParameterSet, see the Thermodynamics for more details

q PhasePartition. Without this argument, the results are for dry air.

The function returns a tuple of

R_m gas_constant_air

cp_m cp_m

cv_m cv_m

γ_m = cp_m/cv_m

(R_m, cp_m, cv_m, γ_m) = gas_constants(ts::ThermodynamicState)

Wrapper to compute all gas constants at once, given a thermodynamic state ts.

The function returns a tuple of

R_m gas_constant_air

cp_m cp_m

cv_m cv_m

γ_m = cp_m/cv_m

ClimateMachine.Thermodynamics.has_condensate — Function

has_condensate(q::PhasePartition{FT})
has_condensate(ts::ThermodynamicState)

Bool indicating if condensate exists in the phase partition

ClimateMachine.Thermodynamics.Ice — Type

Ice <: Phase

An ice phase, to dispatch over saturation_vapor_pressure and q_vap_saturation_generic.

ClimateMachine.Thermodynamics.ice_specific_humidity — Function

ice_specific_humidity(ts::ThermodynamicState)
ice_specific_humidity(q::PhasePartition)

Ice specific humidity given

ts a thermodynamic state

or

q a PhasePartition

ClimateMachine.Thermodynamics.internal_energy — Function

internal_energy(param_set, T[, q::PhasePartition])

The internal energy per unit mass, given a thermodynamic state ts or

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

internal_energy(ts::ThermodynamicState)

The internal energy per unit mass, given a thermodynamic state ts.

internal_energy(ρ::FT, ρe::FT, ρu::AbstractVector{FT}, e_pot::FT)

The internal energy per unit mass, given

ρ (moist-)air density

ρe total energy per unit volume

ρu momentum vector

e_pot potential energy (e.g., gravitational) per unit mass

ClimateMachine.Thermodynamics.internal_energy_dry — Function

internal_energy_dry(param_set, T)

The dry air internal energy

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

internal_energy_dry(ts::ThermodynamicState)

The the dry air internal energy, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.internal_energy_vapor — Function

internal_energy_vapor(param_set, T)

The water vapor internal energy

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

internal_energy_vapor(ts::ThermodynamicState)

The the water vapor internal energy, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.internal_energy_liquid — Function

internal_energy_liquid(param_set, T)

The liquid water internal energy

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

internal_energy_liquid(ts::ThermodynamicState)

The the liquid water internal energy, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.internal_energy_ice — Function

internal_energy_ice(param_set, T)

The ice internal energy

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

internal_energy_ice(ts::ThermodynamicState)

The the ice internal energy, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.internal_energy_sat — Function

internal_energy_sat(param_set, T, ρ, q_tot, phase_type)

The internal energy per unit mass in thermodynamic equilibrium at saturation where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

q_tot total specific humidity

phase_type a thermodynamic state type

internal_energy_sat(ts::ThermodynamicState)

The internal energy per unit mass in thermodynamic equilibrium at saturation, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.latent_heat_fusion — Function

latent_heat_fusion(param_set, T::FT) where {FT<:Real}

The specific latent heat of fusion where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

latent_heat_fusion(ts::ThermodynamicState)

The specific latent heat of fusion given a thermodynamic state ts.

ClimateMachine.Thermodynamics.latent_heat_liq_ice — Function

latent_heat_liq_ice(param_set, q::PhasePartition{FT})

Effective latent heat of condensate (weighted sum of liquid and ice), with specific latent heat evaluated at reference temperature T_0 given

param_set an AbstractParameterSet, see the Thermodynamics for more details

q PhasePartition. Without this argument, the results are for dry air.

ClimateMachine.Thermodynamics.latent_heat_sublim — Function

latent_heat_sublim(param_set, T::FT) where {FT<:Real}

The specific latent heat of sublimation where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

latent_heat_sublim(ts::ThermodynamicState)

The specific latent heat of sublimation given a thermodynamic state ts.

ClimateMachine.Thermodynamics.latent_heat_vapor — Function

latent_heat_vapor(param_set, T::FT) where {FT<:Real}

The specific latent heat of vaporization where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

latent_heat_vapor(ts::ThermodynamicState)

The specific latent heat of vaporization given a thermodynamic state ts.

ClimateMachine.Thermodynamics.Liquid — Type

Liquid <: Phase

A liquid phase, to dispatch over saturation_vapor_pressure and q_vap_saturation_generic.

ClimateMachine.Thermodynamics.liquid_fraction — Function

liquid_fraction(param_set, T, phase_type[, q])

The fraction of condensate that is liquid where

param_set an AbstractParameterSet, see the Thermodynamics for more details

phase_type a thermodynamic state type

PhaseNonEquil behavior

If q.liq or q.ice are nonzero, the liquid fraction is computed from them.

ThermodynamicState

Otherwise, phase equilibrium is assumed so that the fraction of liquid is a function that is 1 above T_freeze and goes to zero below T_freeze.

liquid_fraction(ts::ThermodynamicState)

The fraction of condensate that is liquid given a thermodynamic state ts.

ClimateMachine.Thermodynamics.liquid_ice_pottemp — Function

liquid_ice_pottemp(param_set, T, ρ, q::PhasePartition)

The liquid-ice potential temperature where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

liquid_ice_pottemp(ts::ThermodynamicState)

The liquid-ice potential temperature, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.liquid_ice_pottemp_sat — Function

liquid_ice_pottemp_sat(param_set, T, ρ, phase_type[, q::PhasePartition])

The saturated liquid ice potential temperature where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

phase_type a thermodynamic state type

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

liquid_ice_pottemp_sat(param_set, T, ρ, phase_type, q_tot)

The saturated liquid ice potential temperature where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

phase_type a thermodynamic state type

q_tot total specific humidity

liquid_ice_pottemp_sat(ts::ThermodynamicState)

The liquid potential temperature given a thermodynamic state ts.

ClimateMachine.Thermodynamics.liquid_specific_humidity — Function

liquid_specific_humidity(ts::ThermodynamicState)
liquid_specific_humidity(q::PhasePartition)

Liquid specific humidity given

ts a thermodynamic state

or

q a PhasePartition

ClimateMachine.Thermodynamics.moist_static_energy — Function

moist_static_energy(ts, e_pot)

Moist static energy, given

ts a thermodynamic state

e_pot potential energy (e.g., gravitational) per unit mass

ClimateMachine.Thermodynamics.q_vap_saturation — Function

q_vap_saturation(param_set, T, ρ, phase_type[, q::PhasePartition])

Compute the saturation specific humidity, given

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

phase_type a thermodynamic state type

and, optionally,

q PhasePartition

If the PhasePartition q is given, the saturation specific humidity is that of a mixture of liquid and ice, computed in a thermodynamically consistent way from the weighted sum of the latent heats of the respective phase transitions (Pressel et al., JAMES, 2015). That is, the saturation vapor pressure and from it the saturation specific humidity are computed from a weighted mean of the latent heats of vaporization and sublimation, with the weights given by the fractions of condensates q.liq/(q.liq + q.ice) and q.ice/(q.liq + q.ice) that are liquid and ice, respectively.

If the PhasePartition q is not given, or has zero liquid and ice specific humidities, the saturation specific humidity is that over a mixture of liquid and ice, with the fraction of liquid given by temperature dependent liquid_fraction(param_set, T, phase_type) and the fraction of ice by the complement 1 - liquid_fraction(param_set, T, phase_type).

q_vap_saturation(ts::ThermodynamicState)

Compute the saturation specific humidity, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.q_vap_saturation_liquid — Function

q_vap_saturation_liquid(ts::ThermodynamicState)

Compute the saturation specific humidity over liquid, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.q_vap_saturation_ice — Function

q_vap_saturation_ice(ts::ThermodynamicState)

Compute the saturation specific humidity over ice, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.q_vap_saturation_generic — Function

q_vap_saturation_generic(param_set, T, ρ[, phase=Liquid()])

Compute the saturation specific humidity over a plane surface of condensate, given

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

and, optionally,

Liquid() indicating condensate is liquid

Ice() indicating condensate is ice

ClimateMachine.Thermodynamics.relative_humidity — Function

relative_humidity(param_set, T, p, phase_type, q::PhasePartition)

The relative humidity, given

param_set an AbstractParameterSet, see the Thermodynamics for more details

p pressure

phase_type a thermodynamic state type

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

relative_humidity(ts::ThermodynamicState)

The relative humidity, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.saturated — Function

saturated(ts::ThermodynamicState)

Boolean indicating if thermodynamic state is saturated.

ClimateMachine.Thermodynamics.saturation_adjustment — Function

saturation_adjustment(
    sat_adjust_method,
    param_set,
    e_int,
    ρ,
    q_tot,
    phase_type,
    maxiter,
    temperature_tol
)

Compute the temperature that is consistent with

sat_adjust_method the numerical method to use. See the Thermodynamics for options.

param_set an AbstractParameterSet, see the Thermodynamics for more details

e_int internal energy

ρ (moist-)air density

q_tot total specific humidity

phase_type a thermodynamic state type

maxiter maximum iterations for non-linear equation solve

temperature_tol temperature tolerance

by finding the root of

e_int - internal_energy_sat(param_set, T, ρ, q_tot, phase_type) = 0

using Newtons method with analytic gradients.

See also saturation_adjustment.

ClimateMachine.Thermodynamics.saturation_excess — Function

saturation_excess(param_set, T, ρ, phase_type, q::PhasePartition)

The saturation excess in equilibrium where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

phase_type a thermodynamic state type

q PhasePartition

The saturation excess is the difference between the total specific humidity q.tot and the saturation specific humidity in equilibrium, and it is defined to be nonzero only if this difference is positive.

saturation_excess(ts::ThermodynamicState)

Compute the saturation excess in equilibrium, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.saturation_vapor_pressure — Function

saturation_vapor_pressure(param_set, T, Liquid())

Return the saturation vapor pressure over a plane liquid surface given

T temperature

param_set an AbstractParameterSet, see the Thermodynamics for more details

saturation_vapor_pressure(param_set, T, Ice())

Return the saturation vapor pressure over a plane ice surface given

T temperature

param_set an AbstractParameterSet, see the Thermodynamics for more details

saturation_vapor_pressure(param_set, T, LH_0, Δcp)

Compute the saturation vapor pressure over a plane surface by integration of the Clausius-Clapeyron relation.

The Clausius-Clapeyron relation

`dlog(p_v_sat)/dT = [LH_0 + Δcp * (T-T_0)]/(R_v*T^2)`

is integrated from the triple point temperature T_triple, using Kirchhoff's relation

`L = LH_0 + Δcp * (T - T_0)`

for the specific latent heat L with constant isobaric specific heats of the phases. The linear dependence of the specific latent heat on temperature T allows analytic integration of the Clausius-Clapeyron relation to obtain the saturation vapor pressure p_v_sat as a function of the triple point pressure press_triple.

ClimateMachine.Thermodynamics.soundspeed_air — Function

soundspeed_air(param_set, T[, q::PhasePartition])

The speed of sound in unstratified air, where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

soundspeed_air(ts::ThermodynamicState)

The speed of sound in unstratified air given a thermodynamic state ts.

ClimateMachine.Thermodynamics.specific_enthalpy — Function

specific_enthalpy(e_int, R_m, T)

Specific enthalpy, given

e_int internal specific energy

R_m gas_constant_air

T air temperature

specific_enthalpy(ts)

Specific enthalpy, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.specific_volume — Function

specific_volume(ts::ThermodynamicState)

The (moist-)air specific volume, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.supersaturation — Function

supersaturation(param_set, q, ρ, T, Liquid())
supersaturation(param_set, q, ρ, T, Ice())
supersaturation(ts, Ice())
supersaturation(ts, Liquid())

param_set - abstract set with earth parameters

q - phase partition

ρ - air density,

T - air temperature

Liquid(), Ice() - liquid or ice phase to dispatch over.

ts thermodynamic state

Returns supersaturation (qv/qv_sat -1) over water or ice.

ClimateMachine.Thermodynamics.total_energy — Function

total_energy(param_set, e_kin, e_pot, T[, q::PhasePartition])

The total energy per unit mass, given

param_set an AbstractParameterSet, see the Thermodynamics for more details

e_kin kinetic energy per unit mass

e_pot potential energy per unit mass

T temperature

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

total_energy(e_kin, e_pot, ts::ThermodynamicState)

The total energy per unit mass given a thermodynamic state ts.

ClimateMachine.Thermodynamics.total_specific_enthalpy — Function

total_specific_enthalpy(e_tot, R_m, T)

Total specific enthalpy, given

e_tot total specific energy

R_m gas_constant_air

T air temperature

total_specific_enthalpy(ts)

Total specific enthalpy, given

e_tot total specific energy

ts a thermodynamic state

ClimateMachine.Thermodynamics.total_specific_humidity — Function

total_specific_humidity(ts::ThermodynamicState)
total_specific_humidity(param_set, T, p, relative_humidity)

Total specific humidity given

ts a thermodynamic state

or

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

p pressure

relative_humidity relative humidity (can exceed 1 when there is super saturation/condensate)

ClimateMachine.Thermodynamics.vapor_specific_humidity — Function

vapor_specific_humidity(q::PhasePartition{FT})

The vapor specific humidity, given a PhasePartition q.

ClimateMachine.Thermodynamics.virtual_pottemp — Function

virtual_pottemp(param_set, T, ρ[, q::PhasePartition])

The virtual potential temperature where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

virtual_pottemp(ts::ThermodynamicState)

The virtual potential temperature, given a thermodynamic state ts.

ClimateMachine.Thermodynamics.virtual_temperature — Function

virtual_temperature(param_set, T, ρ[, q::PhasePartition])

The virtual temperature where

param_set an AbstractParameterSet, see the Thermodynamics for more details

T temperature

ρ (moist-)air density

and, optionally,

q PhasePartition. Without this argument, the results are for dry air.

virtual_temperature(ts::ThermodynamicState)

The virtual temperature, given a thermodynamic state ts.

Dispatch types

ClimateMachine.Thermodynamics.DryAdiabaticProcess — Type

DryAdiabaticProcess

For dispatching to isentropic formulas