ODESolvers

LowStorageRungeKutta

CLIMA.ODESolvers.LowStorageRungeKutta2N — Type

LowStorageRungeKutta2N(f, RKA, RKB, RKC, Q; dt, t0 = 0)

This is a time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

\[ \dot{Q} = f(Q, t)\]

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

The constructor builds a low-storage Runge-Kutta scheme using 2N storage based on the provided RKA, RKB and RKC coefficient arrays.

The available concrete implementations are:

LSRK54CarpenterKennedy

LSRK144NiegemannDiehlBusch

CLIMA.ODESolvers.LSRK54CarpenterKennedy — Function

LSRK54CarpenterKennedy(f, Q; dt, t0 = 0)

This function returns a LowStorageRungeKutta2N time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

\[ \dot{Q} = f(Q, t)\]

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

This uses the fourth-order, low-storage, Runge–Kutta scheme of Carpenter and Kennedy (1994) (in their notation (5,4) 2N-Storage RK scheme).

References

@TECHREPORT{CarpenterKennedy1994,
  author = {M.~H. Carpenter and C.~A. Kennedy},
  title = {Fourth-order {2N-storage} {Runge-Kutta} schemes},
  institution = {National Aeronautics and Space Administration},
  year = {1994},
  number = {NASA TM-109112},
  address = {Langley Research Center, Hampton, VA},
}

CLIMA.ODESolvers.LSRK144NiegemannDiehlBusch — Function

LSRK144NiegemannDiehlBusch((f, Q; dt, t0 = 0)

This function returns a LowStorageRungeKutta2N time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

\[ \dot{Q} = f(Q, t)\]

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

This uses the fourth-order, 14-stage, low-storage, Runge–Kutta scheme of Niegemann, Diehl, and Busch (2012) with optimized stability region

References

@article{niegemann2012efficient,
  title={Efficient low-storage Runge--Kutta schemes with optimized stability regions},
  author={Niegemann, Jens and Diehl, Richard and Busch, Kurt},
  journal={Journal of Computational Physics},
  volume={231},
  number={2},
  pages={364--372},
  year={2012},
  publisher={Elsevier}
}

StrongStabilityPreservingRungeKutta

CLIMA.ODESolvers.StrongStabilityPreservingRungeKutta — Type

StrongStabilityPreservingRungeKutta(f, RKA, RKB, RKC, Q; dt, t0 = 0)

This is a time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

\[ \dot{Q} = f(Q, t)\]

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

The constructor builds a strong-stability-preserving Runge–Kutta scheme based on the provided RKA, RKB and RKC coefficient arrays.

The available concrete implementations are:

SSPRK33ShuOsher

SSPRK34SpiteriRuuth

CLIMA.ODESolvers.SSPRK33ShuOsher — Function

SSPRK33ShuOsher(f, Q; dt, t0 = 0)

This function returns a StrongStabilityPreservingRungeKutta time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

\[ \dot{Q} = f(Q, t)\]

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

This uses the third-order, 3-stage, strong-stability-preserving, Runge–Kutta scheme of Shu and Osher (1988)

References

@article{shu1988efficient,
  title={Efficient implementation of essentially non-oscillatory shock-capturing schemes},
  author={Shu, Chi-Wang and Osher, Stanley},
  journal={Journal of computational physics},
  volume={77},
  number={2},
  pages={439--471},
  year={1988},
  publisher={Elsevier}
}

CLIMA.ODESolvers.SSPRK34SpiteriRuuth — Function

SSPRK34SpiteriRuuth(f, Q; dt, t0 = 0)

This function returns a StrongStabilityPreservingRungeKutta time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

\[ \dot{Q} = f(Q, t)\]

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

This uses the third-order, 4-stage, strong-stability-preserving, Runge–Kutta scheme of Spiteri and Ruuth (1988)

References

@article{spiteri2002new,
  title={A new class of optimal high-order strong-stability-preserving time discretization methods},
  author={Spiteri, Raymond J and Ruuth, Steven J},
  journal={SIAM Journal on Numerical Analysis},
  volume={40},
  number={2},
  pages={469--491},
  year={2002},
  publisher={SIAM}
}

AdditiveRungeKutta

CLIMA.ODESolvers.AdditiveRungeKutta — Type

AdditiveRungeKutta(f, l, linearsolver, RKAe, RKAi, RKB, RKC, Q;
                   split_nonlinear_linear, variant, dt, t0 = 0)

This is a time stepping object for implicit-explicit time stepping of a decomposed differential equation. When split_nonlinear_linear == false the equation is assumed to be decomposed as

\[ \dot{Q} = [l(Q, t)] + [f(Q, t) - l(Q, t)]\]

where Q is the state, f is the full tendency and l is the chosen linear operator. When split_nonlinear_linear == true the assumed decomposition is

\[ \dot{Q} = [l(Q, t)] + [f(Q, t)]\]

where f is now only the nonlinear tendency. For both decompositions the linear operator l is integrated implicitly whereas the remaining part is integrated explicitly. Other arguments are the required time step size dt and the optional initial time t0. The resulting linear systems are solved using the provided linearsolver solver. This time stepping object is intended to be passed to the solve! command.

The constructor builds an additive Runge–Kutta scheme based on the provided RKAe, RKAi, RKB and RKC coefficient arrays. Additionally variant specifies which of the analytically equivalent but numerically different formulations of the scheme is used.

The available concrete implementations are:

ARK2GiraldoKellyConstantinescu

ARK548L2SA2KennedyCarpenter

ARK437L2SA1KennedyCarpenter

CLIMA.ODESolvers.ARK2GiraldoKellyConstantinescu — Function

ARK2GiraldoKellyConstantinescu(f, l, linearsolver, Q; dt, t0,
                               split_nonlinear_linear, variant, paperversion)

This function returns an AdditiveRungeKutta time stepping object, see the documentation of AdditiveRungeKutta for arguments definitions. This time stepping object is intended to be passed to the solve! command.

paperversion=true uses the coefficients from the paper, paperversion=false uses coefficients that make the scheme (much) more stable but less accurate

This uses the second-order-accurate 3-stage additive Runge–Kutta scheme of Giraldo, Kelly and Constantinescu (2013).

References

@article{giraldo2013implicit,
  title={Implicit-explicit formulations of a three-dimensional
         nonhydrostatic unified model of the atmosphere ({NUMA})},
  author={Giraldo, Francis X and Kelly, James F and Constantinescu, Emil M},
  journal={SIAM Journal on Scientific Computing},
  volume={35},
  number={5},
  pages={B1162--B1194},
  year={2013},
  publisher={SIAM}
}

CLIMA.ODESolvers.ARK548L2SA2KennedyCarpenter — Function

ARK548L2SA2KennedyCarpenter(f, l, linearsolver, Q; dt, t0,
                            split_nonlinear_linear, variant)

This function returns an AdditiveRungeKutta time stepping object, see the documentation of AdditiveRungeKutta for arguments definitions. This time stepping object is intended to be passed to the solve! command.

This uses the fifth-order-accurate 8-stage additive Runge–Kutta scheme of Kennedy and Carpenter (2013).

References

@article{kennedy2019higher,
  title={Higher-order additive Runge--Kutta schemes for ordinary
         differential equations},
  author={Kennedy, Christopher A and Carpenter, Mark H},
  journal={Applied Numerical Mathematics},
  volume={136},
  pages={183--205},
  year={2019},
  publisher={Elsevier}
}

CLIMA.ODESolvers.ARK437L2SA1KennedyCarpenter — Function

ARK437L2SA1KennedyCarpenter(f, l, linearsolver, Q; dt, t0,
                            split_nonlinear_linear, variant)

This function returns an AdditiveRungeKutta time stepping object, see the documentation of AdditiveRungeKutta for arguments definitions. This time stepping object is intended to be passed to the solve! command.

This uses the fourth-order-accurate 7-stage additive Runge–Kutta scheme of Kennedy and Carpenter (2013).

References

@article{kennedy2019higher,
  title={Higher-order additive Runge--Kutta schemes for ordinary
         differential equations},
  author={Kennedy, Christopher A and Carpenter, Mark H},
  journal={Applied Numerical Mathematics},
  volume={136},
  pages={183--205},
  year={2019},
  publisher={Elsevier}
}

GenericCallbacks

CLIMA.GenericCallbacks — Module

A set of callback functions to be used with an AbstractODESolver

CLIMA.GenericCallbacks.EveryXWallTimeSeconds — Type

EveryXWallTimeSeconds(f, time, mpicomm)

This callback will run the function 'f()' every time wallclock time seconds. The mpicomm is used to syncronize runtime across MPI ranks.

CLIMA.GenericCallbacks.EveryXSimulationSteps — Type

EveryXSimulationSteps(f, steps)

This callback will run the function 'f()' every steps of the time stepper

ODESolvers

CLIMA.ODESolvers.solve! — Function

solve!(Q, solver::AbstractODESolver; timeend,
       stopaftertimeend=true, numberofsteps, callbacks)

Solves an ODE using the solver starting from a state Q. The state Q is updated inplace. The final time timeend or numberofsteps must be specified.

A series of optional callback functions can be specified using the tuple callbacks; see GenericCallbacks.

CLIMA.ODESolvers.gettime — Function

gettime(solver::AbstractODESolver)

Returns the current simulation time of the ODE solver solver

CLIMA.ODESolvers.updatedt! — Function

updatedt!(solver::AbstractODESolver, dt)

Change the time step size to dt for the ODE solver solver.