Ocean Base Module

Hydrostatic Boussinesq

Models

ClimateMachine.Ocean.HydrostaticBoussinesq.HydrostaticBoussinesqModel — Type

HydrostaticBoussinesqModel <: BalanceLaw

A BalanceLaw for ocean modeling.

write out the equations here

ρₒ = reference density of sea water cʰ = maximum horizontal wave speed cᶻ = maximum vertical wave speed αᵀ = thermal expansitivity coefficient νʰ = horizontal viscosity νᶻ = vertical viscosity κʰ = horizontal diffusivity κᶻ = vertical diffusivity fₒ = first coriolis parameter (constant term) β = second coriolis parameter (linear term)

Usage

HydrostaticBoussinesqModel(problem)

ClimateMachine.Ocean.HydrostaticBoussinesq.LinearHBModel — Type

LinearHBModel <: BalanceLaw

A BalanceLaw for modeling vertical diffusion implicitly.

write out the equations here

Usage

model = HydrostaticBoussinesqModel(problem)
linear = LinearHBModel(model)

BCs

ClimateMachine.Ocean.OceanBC — Type

OceanBC(velocity    = Impenetrable(NoSlip())
        temperature = Insulating())

The standard boundary condition for OceanModel. The default options imply a "no flux" boundary condition.

ClimateMachine.Ocean.Penetrable — Type

Penetrable(drag::VelocityDragBC) :: VelocityBC

Defines an penetrable wall model for velocity. This implies:

no constraint on flow in the direction normal to the boundary, and

flow parallel to the boundary is subject to the drag condition.

ClimateMachine.Ocean.Impenetrable — Type

Impenetrable(drag::VelocityDragBC) :: VelocityBC

Defines an impenetrable wall model for velocity. This implies:

no flow in the direction normal to the boundary, and

flow parallel to the boundary is subject to the drag condition.

ClimateMachine.Ocean.NoSlip — Type

NoSlip() :: VelocityDragBC

Zero velocity at the boundary.

ClimateMachine.Ocean.FreeSlip — Type

FreeSlip() :: VelocityDragBC

No surface drag on velocity parallel to the boundary.

ClimateMachine.Ocean.KinematicStress — Type

KinematicStress(stress) :: VelocityDragBC

Applies the specified kinematic stress on velocity normal to the boundary. Prescribe the net inward kinematic stress across the boundary by stress, a function with signature stress(problem, state, aux, t), returning the flux (in m²/s²).

ClimateMachine.Ocean.Insulating — Type

Insulating() :: TemperatureBC

No temperature flux across the boundary

ClimateMachine.Ocean.TemperatureFlux — Type

TemperatureFlux(flux) :: TemperatureBC

Prescribe the net inward temperature flux across the boundary by flux, a function with signature flux(problem, state, aux, t), returning the flux (in m⋅K/s).

ShallowWater

Models

ClimateMachine.Ocean.ShallowWater.ShallowWaterModel — Type

ShallowWaterModel <: BalanceLaw

A BalanceLaw for shallow water modeling.

write out the equations here

Usage

ShallowWaterModel(problem)

OceanProblems

Problems

ClimateMachine.Ocean.OceanProblems.SimpleBox — Type

SimpleBoxProblem <: AbstractSimpleBoxProblem

Stub structure with the dimensions of the box. Lˣ = zonal (east-west) length Lʸ = meridional (north-south) length H = height of the ocean

ClimateMachine.Ocean.OceanProblems.HomogeneousBox — Type

HomogeneousBox <: AbstractSimpleBoxProblem

Container structure for a simple box problem with wind-stress. Lˣ = zonal (east-west) length Lʸ = meridional (north-south) length H = height of the ocean τₒ = maximum value of wind-stress (amplitude)

ClimateMachine.Ocean.OceanProblems.OceanGyre — Type

OceanGyre <: AbstractSimpleBoxProblem

Container structure for a simple box problem with wind-stress, coriolis force, and temperature forcing. Lˣ = zonal (east-west) length Lʸ = meridional (north-south) length H = height of the ocean τₒ = maximum value of wind-stress (amplitude) λʳ = temperature relaxation penetration constant (meters / second) θᴱ = maximum surface temperature

Other (development)

ClimateMachine.Ocean.SuperModels.HydrostaticBoussinesqSuperModel — Type

HydrostaticBoussinesqSuperModel(; domain, time_step, parameters,
     initial_conditions = InitialConditions(),
              advection = (momentum = NonLinearAdvectionTerm(), tracers = NonLinearAdvectionTerm()),
     turbulence_closure = (νʰ=0, νᶻ=0, κʰ=0, κᶻ=0),
               coriolis = (f₀=0, β=0),
    rusanov_wave_speeds = (cʰ=0, cᶻ=0),
               buoyancy = (αᵀ=0,)
       numerical_fluxes = ( first_order = RusanovNumericalFlux(),
                           second_order = CentralNumericalFluxSecondOrder(),
                               gradient = CentralNumericalFluxGradient())
            timestepper = ClimateMachine.ExplicitSolverType(solver_method=LS3NRK33Heuns),
)

Builds a SuperModel that solves the Hydrostatic Boussinesq equations.

ClimateMachine.Ocean.JLD2Writers.JLD2Writer — Type

JLD2Writer(model, outputs=model.fields; filepath, array_type=Array, overwrite_existing=true)

Returns a utility for writing field output to JLD2 files, overwrite_existing file at filepath.

write!(jld2_writer::JLD2Writer) writes outputs to filepath, where outputs is either a NamedTuple or Dictionary of fields or functions of the form output(model).

Field data is converted to array_type before outputting.

ClimateMachine.Ocean.Domains.RectangularDomain — Type

RectangularDomain(FT=Float64;
                  Ne,
                  Np,
                  x = (-1, 1),
                  y = (-1, 1),
                  z = (-1, 1),
                  periodicity = (true, true, false))

Returns a RectangularDomain representing the product of x, y, z intervals, specified by 2-tuples.

The RectangularDomain is meshed with a simple DiscontinuousSpectralElementGrid with an isotropic polynomial order Np and a 3-tuple of Nelements giving the number of elements in x, y, z.

Additional arguments are:

periodicity: a 3-tuple that indicates periodic dimensions with true,

boundary: specifies the boundary condition on each boundary with a boundary condition tag

array_type: either Array for CPU computations or CuArray for GPU computations

mpicomm: communicator for sending data across nodes in a distributed memory configuration using the Message Passing Interface (MPI). See https://pages.tacc.utexas.edu/~eijkhout/pcse/html/mpi-comm.html

Example

julia> using ClimateMachine; ClimateMachine.init()

julia> using ClimateMachine.Ocean.RectangularDomains: RectangularDomain

julia> domain = RectangularDomain(Ne=(7, 8, 9), Np=4, x=(0, 1), y=(0, 1), z=(0, 1))
RectangularDomain{Float64}:
    Np = 4, Ne = (x = 7, y = 8, z = 9)
    L = (x = 1.00e+00, y = 1.00e+00, z = 1.00e+00)
    x = (0.00e+00, 1.00e+00), y = (0.00e+00, 1.00e+00), z = (1.00e+00, 0.00e+00)

ClimateMachine.Ocean.Fields.SpectralElementField — Type

SpectralElementField(domain::RectangularDomain, state::MPIStateArray, variable_index::Int)

Returns a Cartesian view into state.realdata[:, variable_index, :], assuming that state.realdata lives on RectangularDomain.

SpectralElementField.elements is a three-dimensional array of RectangularElements.

SpectralElementField(domain::RectangularDomain, grid, realdata::AbstractArray)

Returns a SpectralElementField whose elements provide a Cartesian view into realdata, assuming that realdata lives on domain::RectangularDomain.

ClimateMachine.Ocean.Fields.RectangularElement — Type

RectangularElement(domain, grid, realdata, element_index)

Returns a Cartesian view into the realdata and nodes (located in grid) associated with element_index.

ClimateMachine.Ocean.Fields.assemble — Function

Assemble an array along the first dimension.

Assemble an array along the second dimension.

Assemble an array along the third dimension.

Assemble elements along dimension.

assemble(elements::Array{<:RectangularElement, 3})

Assemble the three-dimensional data in elements into a single Array, averaging data on shared nodes.

assemble(u::SpectralElementField)

Assemble u.elements into a single element::eltype(u), averaging shared nodes.

assemble_data(u::SpectralElementField{<:CuArray})

Assemble the data in u.elements into a single Array, averaging shared nodes.

ClimateMachine.Ocean.SplitExplicit01.SplitExplicitLSRK2nSolver — Type

SplitExplicitLSRK2nSolver(slow_solver, fast_solver; dt, t0 = 0, coupled = true)

This is a time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

\[ \dot{Q_fast} = f_fast(Q_fast, Q_slow, t) \dot{Q_slow} = f_slow(Q_slow, Q_fast, t)\]

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

This method performs an operator splitting to timestep the Sea-Surface elevation and vertically averaged horizontal velocity of the model at a faster rate than the full model, using LowStorageRungeKutta2N time-stepping.

ClimateMachine.Ocean.SplitExplicit01.SplitExplicitLSRK3nSolver — Type

SplitExplicitLSRK3nSolver(slow_solver, fast_solver; dt, t0 = 0, coupled = true)

This is a time stepping object for explicitly time stepping the differential equation given by the right-hand-side function f with the state Q, i.e.,

\[ \dot{Q_fast} = f_fast(Q_fast, Q_slow, t) \dot{Q_slow} = f_slow(Q_slow, Q_fast, t)\]

with the required time step size dt and optional initial time t0. This time stepping object is intended to be passed to the solve! command.

This method performs an operator splitting to timestep the Sea-Surface elevation and vertically averaged horizontal velocity of the model at a faster rate than the full model, using LowStorageRungeKutta3N time-stepping.

ClimateMachine.Ocean.SplitExplicit01.OceanDGModel — Function

OceanDGModel()

helper function to add required filtering not used in the Driver+Config setup