Single Stack Utils

Functions

ClimateMachine.SingleStackUtils.get_vars_from_nodal_stack — Function

get_vars_from_nodal_stack(
    grid::DiscontinuousSpectralElementGrid{T, dim, Ns},
    Q::MPIStateArray,
    vars;
    vrange::UnitRange = 1:size(Q, 3),
    i::Int = 1,
    j::Int = 1,
    exclude::Vector{String} = String[],
    interp = false,
) where {T, dim, Ns}

Return a dictionary whose keys are the flattenednames() of the variables specified in vars (as returned by e.g. vars_state), and whose values are arrays of the values for that variable along the vertical dimension in Q. Only a single element is expected in the horizontal as this is intended for the single stack configuration and i and j identify the horizontal nodal coordinates.

Variables listed in exclude are skipped.

ClimateMachine.SingleStackUtils.get_vars_from_element_stack — Function

get_vars_from_element_stack(
    grid::DiscontinuousSpectralElementGrid{T, dim, Ns},
    Q::MPIStateArray,
    vars;
    vrange::UnitRange = 1:size(Q, 3),
    exclude::Vector{String} = String[],
    interp = false,
) where {T, dim, Ns}

Return an array of get_vars_from_nodal_stack()s whose dimensions are the number of nodal points per element in the horizontal plane.

Variables listed in exclude are skipped.

ClimateMachine.SingleStackUtils.get_horizontal_mean — Function

get_horizontal_mean(
    grid::DiscontinuousSpectralElementGrid{T, dim, Ns},
    Q::MPIStateArray,
    vars;
    vrange::UnitRange = 1:size(Q, 3),
    exclude::Vector{String} = String[],
    interp = false,
) where {T, dim, Ns}

Return a dictionary whose keys are the flattenednames() of the variables specified in vars (as returned by e.g. vars_state), and whose values are arrays of the horizontal averages for that variable along the vertical dimension in Q. Only a single element is expected in the horizontal as this is intended for the single stack configuration.

Variables listed in exclude are skipped.

ClimateMachine.SingleStackUtils.get_horizontal_variance — Function

get_horizontal_variance(
    grid::DiscontinuousSpectralElementGrid{T, dim, Ns},
    Q::MPIStateArray,
    vars;
    vrange::UnitRange = 1:size(Q, 3),
    exclude::Vector{String} = String[],
    interp = false,
) where {T, dim, Ns}

Return a dictionary whose keys are the flattenednames() of the variables specified in vars (as returned by e.g. vars_state), and whose values are arrays of the horizontal variance for that variable along the vertical dimension in Q. Only a single element is expected in the horizontal as this is intended for the single stack configuration.

Variables listed in exclude are skipped.

ClimateMachine.SingleStackUtils.reduce_nodal_stack — Function

reduce_nodal_stack(
    op::Function,
    grid::DiscontinuousSpectralElementGrid{T, dim, Ns},
    Q::MPIStateArray,
    vars::NamedTuple,
    var::String;
    vrange::UnitRange = 1:size(Q, 3),
) where {T, dim, Ns}

Reduce var from vars within Q over all nodal points in the specified vrange of elements with op. Return a tuple (result, z) where result is the final value returned by op and z is the index within vrange where the result was determined.

ClimateMachine.SingleStackUtils.reduce_element_stack — Function

reduce_element_stack(
    op::Function,
    grid::DiscontinuousSpectralElementGrid{T, dim, Ns},
    Q::MPIStateArray,
    vars::NamedTuple,
    var::String;
    vrange::UnitRange = 1:size(Q, 3),
) where {T, dim, Ns}

Reduce var from vars within Q over all nodal points in the specified vrange of elements with op. Return a tuple (result, z) where result is the final value returned by op and z is the index within vrange where the result was determined.

ClimateMachine.SingleStackUtils.horizontally_average! — Function

horizontally_average!(
    grid::DiscontinuousSpectralElementGrid{T, dim, Ns},
    Q::MPIStateArray,
    i_vars,
) where {T, dim, Ns}

Horizontally average variables, from variable indexes i_vars, in MPIStateArray Q.

Note

These are not proper horizontal averages– the main purpose of this method is to ensure that there are no horizontal fluxes for a single stack configuration.

ClimateMachine.SingleStackUtils.dict_of_nodal_states — Function

dict_of_nodal_states(
    solver_config,
    state_types = (Prognostic(), Auxiliary());
    aux_excludes = [],
    interp = false,
    )

A dictionary of single stack prognostic and auxiliary variables at the i=1,j=1 node given

solver_config a SolverConfiguration

aux_excludes a vector of strings containing the variables to exclude from the auxiliary state.